% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampute.default.R
\name{ampute.default.patterns}
\alias{ampute.default.patterns}
\title{Default `patterns` in `ampute`}
\usage{
ampute.default.patterns(n)
}
\arguments{
\item{n}{A scalar specifying the number of variables in the data.}
}
\value{
A square matrix of size `n` where `0` indicates a variable
}
\description{
This function creates a default pattern matrix for the multivariate
amputation function `ampute()`.
}
\seealso{
[ampute()], [md.pattern()]
}
\author{
Rianne Schouten, 2016
}
\keyword{internal}
