% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmputeDefault.R
\name{ampute.default.weights}
\alias{ampute.default.weights}
\title{Default \code{weights} in \code{ampute}}
\usage{
ampute.default.weights(patterns)
}
\arguments{
\item{patterns}{A matrix of size #patterns by #variables where 0 indicates a 
variable should have missing values and 1 indicates a variable should remain 
complete.}
}
\value{
A matrix of size #patterns by #variables containing the weights that 
will be used to calculate the weighted sum scores. Equal weights are given to 
all variables.
}
\description{
Defines the default weights matrix for the multivariate amputation function 
\code{ampute}.
}
\author{
Rianne Schouten, 2016
}
\seealso{
\code{\link{ampute}}
}

