% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmputeMARCont.R
\name{ampute.mar.cont}
\alias{ampute.mar.cont}
\title{MAR Amputation based on a Continuous probability function}
\usage{
ampute.mar.cont(P, scores, prop, type)
}
\arguments{
\item{P}{A vector containing the values of the patterns the cases are candidate
for. For each case, a value between 1 and #patterns 1 is given. For example, a 
case with value 2 is candidate for missing data pattern 2.}

\item{scores}{A list containing vectors with weighted sum scores of the 
candidates, the result of an underlying function in \code{\link{ampute}}.}

\item{prop}{A scalar specifying the proportion of missingness. Should be value 
between 0 and 1 with 3 decimal places at most. Default is 0.5.}

\item{type}{A vector of strings containing the type of MAR missingness for each 
pattern. Either \code{"MARLEFT"}, \code{"MARMID"}, \code{"MARTAIL"} or 
\code{"MARRIGHT"}. If a single missingness type is entered, all patterns will 
be created by the same type. If missingness types should differ over patterns, 
a vector of missingness types should be entered. Default is MARRIGHT for all 
patterns.}
}
\value{
A list containing vectors with \code{0} if a case should be made missing 
and \code{1} if a case should be kept complete. The first vector refers to the 
first pattern, the second vector to the second pattern, etc.
}
\description{
This function creates a missing data indicator for each pattern. MAR missingness 
is induced by using the probability distributions as described by Van Buuren, 
2012. The function is used in the multivariate amputation function 
\code{\link{ampute}}.
}
\author{
Rianne Schouten, 2016
}
\references{
Van Buuren, S. (2012). \emph{Flexible imputation of missing data} 
(pp. 63-64). Boca Raton, FL.: Chapman & Hall/CRC Press.
}
\seealso{
\code{\link{ampute}}
}

