% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmputeMARDisc.R
\name{ampute.mar.disc}
\alias{ampute.mar.disc}
\title{MAR Amputation based on a Discrete probability function}
\usage{
ampute.mar.disc(P, scores, prop, odds)
}
\arguments{
\item{P}{A vector containing the values of the patterns the cases are candidate
for. For each case, a value between 1 and #patterns 1 is given. For example, a 
case with value 2 is candidate for missing data pattern 2.}

\item{scores}{A list containing vectors with weighted sum scores of the 
candidates, the result of an underlying function in \code{\link{ampute}}.}

\item{prop}{A scalar specifying the proportion of missingness. Should be value 
between 0 and 1 with 3 decimal places at most. Default is 0.5.}

\item{odds}{A matrix where #patterns defines the #rows. Each row should contain 
the odds of being missing for the concurrent pattern. The amount of odds values 
defines in how many quantiles the sum scores will be divided. The values are 
relative probabilities: a quantile with odds value 4 will have a probability of 
being missing that is four times higher than a quantile with odds 1. The 
#quantiles may differ between patterns, specify NA for cells remaining empty. 
Default is 4 quantiles with odds values 1, 2, 3 and 4, for each pattern, 
imitating a MARRIGHT type of missingness.}
}
\value{
A list containing vectors with \code{0} if a case should be made missing 
and \code{1} if a case should be kept complete. The first vector refers to the 
first pattern, the second vector to the second pattern, etc.
}
\description{
This function creates a missing data indicator for each pattern. MAR missingness 
is induced by using odds probabilities as described by Brand, 1999. The function 
is used in the multivariate amputation function \code{\link{ampute}}.
}
\author{
Rianne Schouten, 2016
}
\references{
Brand, J.P.L. (1999). \emph{Development, implementation and 
evaluation of multiple imputation strategies for the statistical analysis of 
incomplete data sets} (pp. 110-113). Dissertation. Rotterdam: Erasmus University.
}
\seealso{
\code{\link{ampute}}
}

