% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmputeMCAR.R
\name{ampute.mcar}
\alias{ampute.mcar}
\title{MCAR Amputation}
\usage{
ampute.mcar(P, patterns, prop)
}
\arguments{
\item{P}{A vector containing the values of the patterns the cases are candidate
for. For each case, a value between 1 and #patterns 1 is given. For example, a 
case with value 2 is candidate for missing data pattern 2.}

\item{patterns}{A matrix of size #patterns by #variables where 0 indicates a variable
should have missing values and 1 indicates a variable should remain complete. 
The user may specify as many patterns as desired. One pattern (a vector) is 
also possible. Default is a square matrix of size #variables where each pattern 
has missingness on one variable only.}

\item{prop}{A scalar specifying the proportion of missingness. Should be value 
between 0 and 1 with 3 decimal places at most. Default is 0.5.}
}
\value{
A list containing vectors with \code{0} if a case should be made missing 
and \code{1} if a case should be kept complete. The first vector refers to the 
first pattern, the second vector to the second pattern, etc.
}
\description{
This function creates a missing data indicator for each pattern, based on MCAR 
missingness. The function is used in the multivariate amputation function 
\code{\link{ampute}}.
}
\author{
Rianne Schouten, 2016
}
\seealso{
\code{\link{ampute}}
}

