% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ampute.R
\name{ampute}
\alias{ampute}
\title{Multivariate Amputation}
\usage{
ampute(data, prop = 0.5, patterns = NULL, freq = NULL,
  mechanism = "MAR", weights = NULL, continuous = TRUE, type = NULL,
  odds = NULL, bycases = TRUE, run = TRUE)
}
\arguments{
\item{data}{A complete data matrix or dataframe. Values should be numeric. 
Categorical variables should have been transformed into dummies.}

\item{prop}{A scalar specifying the proportion of missingness. Should be value 
between 0 and 1 with 3 decimal places at most. Default is 0.5.}

\item{patterns}{A matrix of size #patterns by #variables where 0 indicates a variable
should have missing values and 1 indicates a variable should remain complete. 
The user may specify as many patterns as desired. One pattern (a vector) is 
also possible. Default is a square matrix of size #variables where each pattern 
has missingness on one variable only.}

\item{freq}{A vector of length #patterns containing the relative frequency with 
which the patterns should occur. For example, for three missing data patterns, 
the vector could be \code{c(0.4, 0.4, 0.2)}, meaning that of all cases with 
missing values, 40 percent should have pattern 1, 40 percent pattern 2 and 20 
percent pattern 3. The vector should sum to 1. Default is an equal probability 
for each pattern.}

\item{mechanism}{A string specifying the missingness mechanism, either MCAR or 
MAR. Default is MAR.}

\item{weights}{A matrix of size #patterns by #variables, only useful when 
\code{mechanism = "MAR"}. The matrix contains the weights that will be used
to calculate the weighted sum scores. The weights may differ between patterns 
and between variables. Within each pattern, the relative size of the values are 
of importance. Default is equal weights for all patterns and all variables.}

\item{continuous}{Logical. If TRUE, the probabilities of being missing are based 
on a continuous logit distribution defined by argument \code{type}. If FALSE, the 
probabilities of being missing are based on a discrete distribution defined by 
the \code{odds} argument. Default is TRUE.}

\item{type}{A vector of strings containing the type of MAR missingness for each 
pattern. Either \code{"MARLEFT"}, \code{"MARMID"}, \code{"MARTAIL"} or 
\code{"MARRIGHT"}. If a single missingness type is entered, all patterns will 
be created by the same type. If missingness types should differ over patterns, 
a vector of missingness types should be entered. Default is \code{"MARRIGHT"} for all 
patterns.}

\item{odds}{A matrix where patterns defines the rows. Each row should contain 
the odds of being missing for the concurrent pattern. The amount of odds values 
defines in how many quantiles the sum scores will be divided. The values are 
relative probabilities: a quantile with odds value 4 will have a probability of 
being missing that is four times higher than a quantile with odds 1. The 
quantiles may differ between patterns, specify NA for cells remaining empty. 
Default is 4 quantiles with odds values 1, 2, 3 and 4, for each pattern, 
imitating a MARRIGHT type of missingness.}

\item{bycases}{Logical. If TRUE, the proportion of missingness is defined in 
terms of cases. If FALSE, the proportion of missingness is defined in terms of 
cells. Default is TRUE.}

\item{run}{Logical. If TRUE, the amputations are implemented. If FALSE, the 
return object will contain everything but the amputed data set.}
}
\value{
Returns a S3 object of class \code{\link{mads-class}} (multivariately amputed 
data set)
}
\description{
This function amputes multivariate data in a MCAR or MAR manner. The details 
section gives a concise explanation of the why and how of this function.
}
\details{
When new multiple imputation techniques are tested, values in simulated 
data sets have to be made missing. Often, a univariate amputation procedure is 
followed, and then repeated multiple times in case several variables should
have missing values. However, this approach makes it difficult to relate the 
missingness on one variable to the missingness on another variable. A 
multivariate amputation procedure solves this issue and moreover, it does 
justice to the multivariate nature of data sets. Hence, \code{ampute} is 
created to perform the amputation according the researcher's desires.

The idea behind the function is the specification of several missingness 
patterns. Each pattern is a combination of variables with and without missing 
values (denoted by \code{0} and \code{1} respectively). For example, one might
want to create two missingness patterns on a data set with four variables. The
patterns would be something like: \code{0, 0, 1, 1} and \code{1, 0, 1, 0}. 
Each combination of zeros and ones may occur, except a pattern of merely ones 
or zeros.

Furthermore, the researcher specifies the proportion of missingness, either the 
proportion of missing cases or the proportion of missing cells, and the relative 
frequency each pattern occurs. Consequently, the data is divided over the 
patterns with these probabilities. Now, each case is candidate for a certain 
missingness pattern. Whether the case will have missing values eventually, 
depends on other specifications.

When the user sets the missingness mechanism to \code{"MCAR"}, the candidates 
have an equal probability of having missing values. No other specifications 
have to be made. For a \code{"MAR"} mechanism, weighted sum scores are calculated. 
First, the data is standardized. Second, for each case, the observed values are 
multiplied with weights, specified by argument \code{weights}. These weighted
scores will be summed, resulting in a weighted sum score for each case.

The weights may differ between patterns and they may be negative or zero as well. 
Naturally, it is the relative difference between the weights that will result
in a \code{MAR} effect in the sum scores. For example, for the first missing data
pattern mentioned above, the weights for the third and fourth variables might
be set to 2 and 4. However, weight values of 0.2 and 0.4 will have the exact
same effect on the weighted sum score: the fourth variable is weighted twice as
much as variable 3. In this specific pattern, variable 1 and 2 will be set missing. 
Consequently, their weights are not of importance.

Based on the weighted sum scores, either a discrete or continuous distribution 
of probabilities is used to calculate whether a candidate will have missing values.

For a discrete distribution of probabilities, the weighted sum scores are 
divided into subgroups of equal size (quantiles). Thereafter, the user 
specifies for each subgroup the odds of being missing. Both the number of 
subgroups and the odds values guide the incomplete data mechanism for a certain 
pattern. For example, for a MARRIGHT-like mechanism, scoring in one of the 
higher quantiles should have high missingness odds, whereas for a MARMID-like 
mechanism, the central groups should have higher odds. Again, not the size of 
the odds values are of importance, but the relative distance between the values.

The continuous distribution of probabilities is based on the logit function, as 
described by Van Buuren (2012). The user can specify the type of missingness, 
which, again, may differ between patterns.

An extensive example is shown in a vignette 
titled "Multivariate Amputation using Ampute".
}
\examples{
# Simulate data set with \\code{mvrnorm} from package \\code{\\pkg{MASS}}.
require(MASS)
sigma <- matrix(data = c(1, 0.2, 0.2, 0.2, 1, 0.2, 0.2, 0.2, 1), nrow = 3)
complete.data <- mvrnorm(n = 100, mu = c(5, 5, 5), Sigma = sigma)
# Perform quick amputation
result1 <- ampute(data = complete.data)
# Change default matrices as desired
patterns <- result1$patterns
patterns[1:3, 2] <- 0
odds <- result1$odds
odds[2,3:4] <- c(2, 4)
odds[3,] <- c(3, 1, NA, NA)
# Rerun amputation
result3 <- ampute(data = complete.data, patterns = patterns, freq = 
c(0.3, 0.3, 0.4), continuous = FALSE, odds = odds)
# Run an amputation procedure with continuous probabilities
result4 <- ampute(data = complete.data, type = c("MARRIGHT", "MARTAIL", 
"MARLEFT"))

}
\author{
Rianne Schouten, Gerko Vink, Peter Lugtig, 2016
}
\references{
Brand, J.P.L. (1999). \emph{Development, implementation and 
evaluation of multiple imputation strategies for the statistical analysis of 
incomplete data sets} (pp. 110-113). Dissertation. Rotterdam: Erasmus University. 

Van Buuren, S., Brand, J.P.L., Groothuis-Oudshoorn, C.G.M., Rubin, D.B. (2006). 
Fully conditional specification in multivariate imputation. \emph{Journal of 
Statistical Computation and Simulation}, 76\emph{(12)}, Appendix B.

Van Buuren, S. (2012). \emph{Flexible imputation of missing data} (pp. 63-64). 
Boca Raton, FL.: Chapman & Hall/CRC Press.

Vink, G. (2016). Towards a standardized evaluation of multiple imputation 
routines.
}
\seealso{
\code{\link{mads-class}}
}

