% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mira}
\alias{anova.mira}
\title{Compare several nested models}
\usage{
\method{anova}{mira}(object, ..., method = "D1", use = "wald")
}
\arguments{
\item{object}{Two or more objects of class `mira`}

\item{...}{Other parameters passed down to `D1()`, `D2()`,
`D3()` and `mitml::testModels`.}

\item{method}{Either `"D1"`, `"D2"` or `"D3"`}

\item{use}{An character indicating the test statistic}
}
\value{
Object of class `mice.anova`
}
\description{
Compare several nested models
}
