% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mira}
\alias{as.mira}
\title{Create a `mira` object from repeated analyses}
\usage{
as.mira(fitlist)
}
\arguments{
\item{fitlist}{A list containing $m$ fitted analysis objects}
}
\value{
An S3 object of class `mira`.
}
\description{
The `as.mira()` function takes the results of repeated
complete-data analysis stored as a list, and turns it
into a `mira` object that can be pooled.
}
\seealso{
[`mira()`][mira-class]
}
\author{
Stef van Buuren
}
