% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.coef.R
\name{assign.to.global}
\alias{assign.to.global}
\title{function loading results in global environment}
\usage{
assign.to.global(x, value, pos = 1)
}
\arguments{
\item{x}{a variable name, given as a character string.  No coercion is
    done, and the first element of a character vector of length greater
    than one will be used, with a warning.}

\item{value}{a value to be assigned to \code{x}.}

\item{pos}{where to do the assignment.  By default, assigns into the
    current environment.  See \sQuote{Details} for other possibilities.}
}
\description{
function loading results in global environment
}
\note{
This function will silence R CMD CHECK 
"Found the following assignments to the global environment:"
}
\keyword{internal}
