% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandsma.R
\docType{data}
\name{brandsma}
\alias{brandsma}
\title{Brandsma school data used Snijders and Bosker (2012)}
\format{
`brandsma` is a data frame with 4106 rows and 14 columns:
\describe{
\item{`sch`}{School number}
\item{`pup`}{Pupil ID}
\item{`iqv`}{IQ verbal}
\item{`iqp`}{IQ performal}
\item{`sex`}{Sex of pupil}
\item{`ses`}{SES score of pupil}
\item{`min`}{Minority member 0/1}
\item{`rpg`}{Number of repeated groups, 0, 1, 2}
\item{`lpr`}{language score PRE}
\item{`lpo`}{language score POST}
\item{`apr`}{Arithmetic score PRE}
\item{`apo`}{Arithmetic score POST}
\item{`den`}{Denomination classification 1-4 - at school level}
\item{`ssi`}{School SES indicator - at school level}
}
}
\source{
Constructed from `MLbook_2nded_total_4106-99.sav` from
<https://www.stats.ox.ac.uk/~snijders/mlbook.htm> by function
`data-raw/R/brandsma.R`
}
\description{
Dataset with raw data from Snijders and Bosker (2012) containing
data from 4106 pupils attending 216 schools. This dataset
includes all pupils and schools with missing data.
}
\note{
This dataset is constructed from the raw data. There are
a few differences with the data set used in Chapter 4 and 5
of Snijders and Bosker:
\enumerate{
\item All schools are included, including the five school with
missing values on `langpost`.
\item Missing `denomina` codes are left as missing.
\item Aggregates are undefined in the presence of missing data
in the underlying values.
Variables `ses`, `iqv` and `iqp` are in their
original scale, and not globally centered.
No aggregate variables at the school level are included.
\item There is a wider selection of original variables. Note
however that the source data contain an even wider set of
variables.
}
}
\references{
Brandsma, HP and Knuver, JWM (1989), Effects of school and
classroom characteristics on pupil progress in language and arithmetic.
International Journal of Educational Research, 13(7), 777 - 788.

Snijders, TAB and Bosker RJ (2012). Multilevel Analysis, 2nd Ed. Sage,
Los Angeles, 2012.
}
\keyword{datasets}
