% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.mads.R
\name{bwplot.mads}
\alias{bwplot.mads}
\title{Box-and-whisker plot of amputed and non-amputed data}
\usage{
\method{bwplot}{mads}(x, yvar = NULL, which.pat = NULL,
  standardized = TRUE, descriptives = TRUE, layout = NULL,
  theme = mice.theme(), multiple = TRUE, outer = TRUE)
}
\arguments{
\item{x}{A \code{mads} object, typically created by \code{\link{ampute}}.}

\item{yvar}{A string or vector of variable names that needs to be plotted. As 
a default, all variables will be plotted.}

\item{which.pat}{A scalar or vector indicating which patterns need to be plotted. 
As a default, all patterns are plotted.}

\item{standardized}{Logical. Whether the box-and-whisker plots need to be created
from standardized data or not. Default is TRUE.}

\item{descriptives}{Logical. Whether the mean, variance and n of the variables
need to be printed. This is useful to examine the effect of the amputation. 
Default is TRUE.}

\item{layout}{A vector of two values indicating how the boxplots of one pattern
should be divided over the plot. For example, \code{c(2, 3)} indicates that the 
boxplots of six variables need to be placed on 3 rows and 2 columns. Default
is 1 row and an amount of columns equal to #variables. Note that for more than 
6 variables, multiple plots will be created automatically.}

\item{theme}{A named list containing the graphical parameters. The default
function \code{mice.theme} produces a short list of default colors, line
width, and so on. The extensive list may be obtained from
\code{trellis.par.get}.}

\item{multiple}{See \code{\link[lattice:bwplot]{bwplot}}.}

\item{outer}{See \code{\link[lattice:bwplot]{bwplot}}.}
}
\value{
A list containing the box-and-whisker plots. Note that a new pattern 
will always be shown in a new plot.
}
\description{
Plotting method for amputed data using \pkg{lattice}. \code{bwplot}
produces box-and-whisker plots. The function automatically separates the 
amputed and non-amputed data.
}
\note{
The \code{mads} object contains all the information you need to 
make any desired plots. Check \code{\link{mads-class}} or the vignette titled "Multivariate 
Amputation using Ampute" to understand the contents of class object \code{mads}.
}
\author{
Rianne Schouten, 2016
}
\seealso{
\code{\link{ampute}}, \code{\link{bwplot}}, \code{\link{Lattice}} for 
an overview of the package, \code{\link{mads-class}}
}

