% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.r
\name{cbind}
\alias{cbind}
\title{Combine objects by columns}
\usage{
cbind(...)
}
\arguments{
\item{...}{(generalized) vectors or matrices.  These can be given as named
    arguments.  Other \R objects may be coerced as appropriate, or S4
    methods may be used: see sections \sQuote{Details} and
    \sQuote{Value}.  (For the \code{"data.frame"} method of \code{cbind}
    these can be further arguments to \code{\link{data.frame}} such as
    \code{stringsAsFactors}.)}
}
\description{
Overwrite the base::cbind function to enable proper dispatch for data.frames
https://stackoverflow.com/questions/34024585/s3-dispatching-of-rbind-and-cbind
}
\keyword{internal}
