% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Select complete cases}
\usage{
cc(x)
}
\arguments{
\item{x}{An `R` object. Methods are available for classes
`mids`, `data.frame` and `matrix`. Also, `x`
could be a vector.}
}
\value{
A `vector`, `matrix` or `data.frame` containing the data of the complete cases.
}
\description{
Extracts the complete cases, also known as *listwise deletion*.
`cc(x)` is similar to
`na.omit(x)`, but returns an object of the same class
as the input data. Dimensions are not dropped. For extracting
incomplete cases, use [ici()].
}
\examples{

# cc(nhanes)   # get the 13 complete cases
# cc(nhanes$bmi) # extract complete bmi
}
\seealso{
[na.omit()], [cci()], [ici()]
}
\author{
Stef van Buuren, 2017.
}
\keyword{univar}
