% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{cci}
\alias{cci}
\title{Complete case indicator}
\usage{
cci(x)
}
\arguments{
\item{x}{An `R` object. Currently supported are methods for the
following classes: `mids`.}
}
\value{
Logical vector indicating the complete cases.
}
\description{
The complete case indicator is useful for extracting the subset of complete cases. The function
`cci(x)` calls `complete.cases(x)`.
The companion function `ici()` selects the incomplete cases.
}
\examples{
cci(nhanes) # indicator for 13 complete cases
cci(mice(nhanes, maxit = 0))
f <- cci(nhanes[, c("bmi", "hyp")]) # complete data for bmi and hyp
nhanes[f, ] # obtain all data from those with complete bmi and hyp
}
\seealso{
[complete.cases()], [ici()], [cc()]
}
\author{
Stef van Buuren, 2017.
}
\keyword{univar}
