% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coerce.newdata}
\alias{coerce.newdata}
\title{Coerce a data.frame to match model expectations}
\usage{
coerce.newdata(data, models)
}
\arguments{
\item{data}{A \code{data.frame} with input data.}

\item{models}{A list of trained models from \code{mice()}.}
}
\value{
A coerced \code{data.frame} that matches the class and levels from \code{models}.
}
\description{
This function coerces columns in \code{data} to match the type and levels
expected by the corresponding trained \code{models}.
}
