% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{create.formula}
\alias{create.formula}
\title{Creates a formula that combines form and predictors}
\usage{
create.formula(form = ~0, predictors = NULL, include.predictors = TRUE,
  include.intercept = FALSE, ...)
}
\arguments{
\item{form}{A formula, typically specified by the user. If it is 
not a formula, the form is internally reset to \code{~0}.}

\item{predictors}{A character vector of variable names.}

\item{include.predictors}{A logical that indicates whether the variables
listed in \code{predictors} should be added to the result. The default 
is \code{TRUE}}

\item{include.intercept}{A logical that indicated whether the intercept 
should be included in the result. The default (\code{FALSE}) may change 
in future versions.}
}
\value{
A formula
}
\description{
Creates a formula that combines form and predictors
}
\keyword{internal}
