% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1.R
\name{D1}
\alias{D1}
\title{Compare two nested models using D1-statistic}
\usage{
D1(model, null.model = NULL, ...)
}
\arguments{
\item{model}{An object of class 'mira', produced by \code{with.mids()}.}

\item{null.model}{An object of class 'mira', produced by \code{with.mids()}. The
model in \code{fit0} is a nested fit0 of \code{fit1}.}

\item{\dots}{Passed down to \code{\link[mitml]{testModels}}}
}
\description{
Compare two nested models using D1-statistic
}
