% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2.R
\name{D2}
\alias{D2}
\title{Compare two nested models using D2-statistic}
\usage{
D2(fit1, fit0 = NULL, ...)
}
\arguments{
\item{fit1}{An object of class \code{mira}, produced by \code{with()}.}

\item{fit0}{An object of class \code{mira}, produced by \code{with()}. The
model in \code{fit0} is a nested within \code{fit1}. The default null 
model \code{fit0 = NULL} compares \code{fit1} to the intercept-only model.}

\item{...}{Not used.}
}
\description{
Compare two nested models using D2-statistic
}
