% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3.R
\name{D3}
\alias{D3}
\title{Compare two nested models using D3-statistic}
\usage{
D3(fit1, fit0 = NULL, ...)
}
\arguments{
\item{fit1}{An object of class 'mira', produced by \code{with.mids()}.}

\item{fit0}{An object of class 'mira', produced by \code{with.mids()}. The
model in \code{fit0} is a nested fit0 of \code{fit1}.}

\item{\dots}{Not used}
}
\description{
Compare two nested models using D3-statistic
}
\author{
Stef van Buuren, 2018
}
