% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.spcr.R
\name{.spcrCVE}
\alias{.spcrCVE}
\title{Computes the cross-validation error}
\usage{
.spcrCVE(dv, pred, part, K = 10, npcs = 1)
}
\arguments{
\item{dv}{Vector of observed values on the dv.}

\item{pred}{Matrix of complete covariates for a version of a model.}

\item{part}{Vector indicating how to partition data in K-folds.}

\item{K}{Unit vector (double) indicating the number of folds for cross-validation.}

\item{npcs}{Unit vector (double) indicating the number of components to retain.}
}
\description{
Computes the cross-validation error
}
