% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{extend.formula}
\alias{extend.formula}
\title{Extends a formula with predictors}
\usage{
extend.formula(formula = ~0, predictors = NULL, auxiliary = TRUE,
  include.intercept = FALSE, ...)
}
\arguments{
\item{formula}{A formula. If it is 
not a formula, the formula is internally reset to \code{~0}.}

\item{predictors}{A character vector of variable names.}

\item{auxiliary}{A logical that indicates whether the variables
listed in \code{predictors} should be added to the formula as main 
effects. The default is \code{TRUE}.}

\item{include.intercept}{A logical that indicated whether the intercept 
should be included in the result.}
}
\value{
A formula
}
\description{
Extends a formula with predictors
}
\keyword{internal}
