% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{extend.formulas}
\alias{extend.formulas}
\title{Extends formula's with predictor matrix settings}
\usage{
extend.formulas(
  formulas,
  data,
  blocks,
  predictorMatrix = NULL,
  auxiliary = TRUE,
  include.intercept = FALSE,
  ...
)
}
\arguments{
\item{formulas}{A named list of formula's, or expressions that
can be converted into formula's by \code{as.formula}. List elements
correspond to blocks. The block to which the list element applies is
identified by its name, so list names must correspond to block names.
The \code{formulas} argument is an alternative to the
\code{predictorMatrix} argument that allows for more flexibility in
specifying imputation models, e.g., for specifying interaction terms.}

\item{data}{Data frame with \eqn{n} rows and \eqn{p} columns with
incomplete data.  Missing values are coded as \code{NA}.}

\item{blocks}{List of vectors with variable names per block. List elements
may be named to identify blocks. Variables within a block are
imputed by a multivariate imputation method
(see \code{method} argument). By default each variable is placed
into its own block, which is effectively
fully conditional specification (FCS) by univariate models
(variable-by-variable imputation). Only variables whose names appear in
\code{blocks} are imputed. The relevant columns in the \code{where}
matrix are set to \code{FALSE} of variables that are not block members.
A variable may appear in multiple blocks. In that case, it is
effectively re-imputed each time that it is visited.}

\item{predictorMatrix}{A square numeric matrix of \eqn{p} rows
and columns. Row- and column names are \code{colnames(data)}.
Each row corresponds to a variable to be imputed.
A value of \code{1} means that the column variable is a
predictor for the row variable, while a \code{0} means that
the column variable is not a predictor. The default
\code{predictorMatrix} is \code{1} everywhere, except for a zero
diagonal. For variables that need no be imputed,
\code{mice()} automatically sets the corresponding rows in the
\code{predictorMatrix} to zero. See details
on \emph{skipping imputation}.
Two-level imputation models (which have \code{"2l"} in their
names) other codes than \code{0} and \code{1}, e.g, \code{2} or \code{-2},
are also used.}

\item{auxiliary}{A logical that indicates whether the variables
listed in \code{predictors} should be added to the formula as main
effects. The default is \code{TRUE}.}

\item{include.intercept}{A logical that indicated whether the intercept
should be included in the result.}

\item{...}{Named arguments that are passed down to the univariate imputation
functions.}
}
\value{
A list of formula's
}
\description{
Extends formula's with predictor matrix settings
}
\keyword{internal}
