% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_imputations.R
\name{extract_imputations}
\alias{extract_imputations}
\title{Extract mice imputations}
\usage{
extract_imputations(miss_dat, mice, j = 1)
}
\arguments{
\item{miss_dat}{the dataset containing missing values}

\item{mice}{the "imp" object from the 'mids-class' object returned by mice
eg. mice_obj$imp}

\item{j}{For multiple imputations, which imputation to extract. Defaults to 1.}
}
\value{
A an equivalent dataset to \code{miss_dat} with imputed observations
}
\description{
Extract the imputed observations from a 'mids-class' object for a 
given multiple imputation
}

