% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{extractBS}
\alias{extractBS}
\title{Extract broken stick estimates from a `lmer` object}
\usage{
extractBS(fit)
}
\arguments{
\item{fit}{An object of class `lmer`}
}
\value{
A matrix containing broken stick estimates
}
\description{
Extract broken stick estimates from a `lmer` object
}
\author{
Stef van Buuren, 2012
}
