% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{fico}
\alias{fico}
\title{Fraction of incomplete cases among cases with observed}
\usage{
fico(data)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as NA's.}
}
\value{
A vector of length `ncol(data)` of FICO statistics.
}
\description{
FICO is an outbound statistic defined by the fraction of incomplete cases
among cases with `Yj` observed (White and Carlin, 2010).
}
\references{
Van Buuren, S. (2018).
[*Flexible Imputation of Missing Data. Second Edition.*](https://stefvanbuuren.name/fimd/missing-data-pattern.html#sec:flux)
Chapman & Hall/CRC. Boca Raton, FL.

White, I.R., Carlin, J.B. (2010). Bias and efficiency of multiple imputation
compared with complete-case analysis for missing covariate values.
*Statistics in Medicine*, *29*, 2920-2931.
}
\seealso{
[fluxplot()], [flux()], [md.pattern()]
}
\author{
Stef van Buuren, 2012
}
\keyword{misc}
