% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.coef.R
\name{fix.coef}
\alias{fix.coef}
\title{Fix coefficients and update model}
\usage{
fix.coef(model, beta = NULL)
}
\arguments{
\item{model}{An R model, e.g., produced by \code{lm} or \code{glm}}

\item{beta}{A numeric vector with \code{length(coef)} regression weights}
}
\value{
An updated R model object
}
\description{
Refits a model with a specified set of coefficients.
}
\examples{
fit0 <- lm(Volume ~ Girth + Height, data = trees)
coef(fit0)
deviance(fit0)

# refit same model
fit1 <- fix.coef(fit0)
coef(fit1)
deviance(fit1)

# change the beta's
fit2 <- fix.coef(fit0, beta = c(-50, 5, 1))
coef(fit2)
deviance(fit2)

plot(predict(fit0), predict(fit1)); abline(0,1)
plot(predict(fit0), predict(fit2)); abline(0,1)

\dontrun{
# it fails on missing data
fit0 <- lm(bmi ~ age + hyp + chl, data = nhanes2)
fit1 <- fix.coef(fit0)
}
}
\author{
Stef van Buuren
}
