% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{fluxplot}
\alias{fluxplot}
\title{Fluxplot of the missing data pattern}
\usage{
fluxplot(
  data,
  local = names(data),
  plot = TRUE,
  labels = TRUE,
  xlim = c(0, 1),
  ylim = c(0, 1),
  las = 1,
  xlab = "Influx",
  ylab = "Outflux",
  main = paste("Influx-outflux pattern for", deparse(substitute(data))),
  eqscplot = TRUE,
  pty = "s",
  lwd = 1,
  ...
)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as NA's.}

\item{local}{A vector of names of columns of `data`. The default is to
include all columns in the calculations.}

\item{plot}{Should a graph be produced?}

\item{labels}{Should the points be labeled?}

\item{xlim}{See `par`.}

\item{ylim}{See `par`.}

\item{las}{See `par`.}

\item{xlab}{See `par`.}

\item{ylab}{See `par`.}

\item{main}{See `par`.}

\item{eqscplot}{Should a square plot be produced?}

\item{pty}{See `par`.}

\item{lwd}{See `par`. Controls axis line thickness and diagonal}

\item{\dots}{Further arguments passed to `plot()` or `eqscplot()`.}
}
\value{
An invisible data frame with `ncol(data)` rows and six columns:
pobs = Proportion observed,
influx = Influx
outflux = Outflux
ainb = Average inbound statistic
aout = Average outbound statistic
fico = Fraction of incomplete cases among cases with `Yj` observed
}
\description{
Influx and outflux are statistics of the missing data pattern. These
statistics are useful in selecting predictors that should go into the
imputation model.
}
\details{
Infux and outflux have been proposed by Van Buuren (2012), chapter 4.

Influx is equal to the number of variable pairs `(Yj , Yk)` with
`Yj` missing and `Yk` observed, divided by the total number of
observed data cells. Influx depends on the proportion of missing data of the
variable. Influx of a completely observed variable is equal to 0, whereas for
completely missing variables we have influx = 1. For two variables with the
same proportion of missing data, the variable with higher influx is better
connected to the observed data, and might thus be easier to impute.

Outflux is equal to the number of variable pairs with `Yj` observed and
`Yk` missing, divided by the total number of incomplete data cells.
Outflux is an indicator of the potential usefulness of `Yj` for imputing
other variables. Outflux depends on the proportion of missing data of the
variable. Outflux of a completely observed variable is equal to 1, whereas
outflux of a completely missing variable is equal to 0. For two variables
having the same proportion of missing data, the variable with higher outflux
is better connected to the missing data, and thus potentially more useful for
imputing other variables.
}
\references{
Van Buuren, S. (2018).
[*Flexible Imputation of Missing Data. Second Edition.*](https://stefvanbuuren.name/fimd/missing-data-pattern.html#sec:flux)
Chapman & Hall/CRC. Boca Raton, FL.

White, I.R., Carlin, J.B. (2010). Bias and efficiency of multiple imputation
compared with complete-case analysis for missing covariate values.
*Statistics in Medicine*, *29*, 2920-2931.
}
\seealso{
[flux()], [md.pattern()], [fico()]
}
\author{
Stef van Buuren, 2012
}
\keyword{misc}
