% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.mira}
\alias{glance.mira}
\title{Glance method to extract information from a `mira` object}
\usage{
\method{glance}{mira}(x, ...)
}
\arguments{
\item{x}{An object with multiply-imputed models from `mice` (class: `mira`)}

\item{...}{extra arguments (not used)}
}
\value{
a dataframe with one row and the following columns:
\itemize{
  \item nimp
  \item nobs
}
}
\description{
Glance method to extract information from a `mira` object
}
\note{
If x contains `lm` models, R2 and Adj.R2 are included in the output
}
\concept{tidiers}
\keyword{internal}
