% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{glm.mids}
\alias{glm.mids}
\title{Generalized linear model for `mids` object}
\usage{
glm.mids(formula, family = gaussian, data, ...)
}
\arguments{
\item{formula}{a formula expression as for other regression models, of the
form response ~ predictors. See the documentation of [lm()] and
[formula()] for details.}

\item{family}{The family of the glm model}

\item{data}{An object of type `mids`, which stands for 'multiply imputed
data set', typically created by function `mice()`.}

\item{\dots}{Additional parameters passed to [glm()].}
}
\value{
An objects of class `mira`, which stands for 'multiply imputed
repeated analysis'.  This object contains `data$m` distinct
`glm.objects`, plus some descriptive information.
}
\description{
Applies `glm()` to a multiply imputed data set
}
\details{
This function is included for backward compatibility with V1.0. The function
is superseded by [with.mids()].
}
\examples{

imp <- mice(nhanes)

# logistic regression on the imputed data
fit <- glm.mids((hyp == 2) ~ bmi + chl, data = imp, family = binomial)
fit
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, C.G.M. (2000)
*Multivariate Imputation by Chained Equations: MICE V1.0 User's manual.*
Leiden: TNO Quality of Life.
}
\seealso{
[with.mids()], [glm()], [`mids()`][mids-class],
[`mira()`][mira-class]
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000
}
\keyword{multivariate}
