% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibind.R
\name{ibind}
\alias{ibind}
\title{Enlarge number of imputations by combining `mids` objects}
\usage{
ibind(x, y)
}
\arguments{
\item{x}{A `mids` object.}

\item{y}{A `mids` object.}
}
\value{
An S3 object of class `mids`
}
\description{
This function combines two `mids` objects `x` and `y` into a
single `mids` object, with the objective of increasing the number of
imputed data sets. If the number of imputations in `x` and `y` are
`m(x)` and `m(y)`, then the combined object will have
`m(x)+m(y)` imputations.
}
\details{
The two `mids` objects are required to
have the same underlying multiple imputation model and should
be fitted on the same data.
}
\examples{
data(nhanes)
imp1 <- mice(nhanes, m = 1, maxit = 2, print = FALSE)
imp1$m

imp2 <- mice(nhanes, m = 3, maxit = 3, print = FALSE)
imp2$m

imp12 <- ibind(imp1, imp2)
imp12$m
plot(imp12)
}
\seealso{
[`mids()`][mids-class]
}
\author{
Karin Groothuis-Oudshoorn, Stef van Buuren
}
\keyword{manip}
