% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{ic}
\alias{ic}
\title{Select incomplete cases}
\usage{
ic(x)
}
\arguments{
\item{x}{An `R` object. Methods are available for classes
`mids`, `data.frame` and `matrix`. Also, `x`
could be a vector.}
}
\value{
A `vector`, `matrix` or `data.frame` containing the data of the complete cases.
}
\description{
Extracts incomplete cases from a data set.
The companion function for selecting the complete cases is [cc()].
}
\examples{

ic(nhanes) # get the 12 rows with incomplete cases
ic(nhanes[1:10, ]) # incomplete cases within the first ten rows
ic(nhanes[, c("bmi", "hyp")]) # restrict extraction to variables bmi and hyp
}
\seealso{
[cc()], [ici()]
}
\author{
Stef van Buuren, 2017.
}
\keyword{univar}
