% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.r
\name{ici}
\alias{ici}
\alias{ici,data.frame-method}
\alias{ici,matrix-method}
\alias{ici,mids-method}
\title{Incomplete case indicator}
\usage{
ici(x)
}
\arguments{
\item{x}{An \code{R} object. Currently supported are methods for the
following classes: \code{mids}.}
}
\value{
Logical vector indicating the incomplete cases,
}
\description{
This array is useful for extracting the subset of incomplete cases.
The companion function  \code{cci()} selects the complete cases.
}
\examples{

 ici(nhanes) # indicator for 12 rows with incomplete cases 

}
\seealso{
\code{\link{cci}}, \code{\link{ic}}
}
\author{
Stef van Buuren, 2017.
}
\keyword{univar}
