\name{ifdo}
\alias{ifdo}
\title{Conditional imputation helper}
\usage{
  ifdo(cond, action)
}
\arguments{
  \item{cond}{}

  \item{action}{}
}
\value{
  Currently returns an error message.
}
\description{
  Sorry, the \code{ifdo()} function is not yet implemented.
}
\author{
  Stef van Buuren, 2012
}
\keyword{internal}

