% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mads}
\alias{is.mads}
\title{Check for `mads` object}
\usage{
is.mads(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether `x` is an object of class `mads`
}
\description{
Check for `mads` object
}
