% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mids}
\alias{is.mids}
\title{Check for `mids` object}
\usage{
is.mids(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether `x` is an object of class `mids`
}
\description{
Check for `mids` object
}
