% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mipo}
\alias{is.mipo}
\title{Check for `mipo` object}
\usage{
is.mipo(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether `x` is an object of class `mipo`
}
\description{
Check for `mipo` object
}
