% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mira}
\alias{is.mira}
\title{Check for `mira` object}
\usage{
is.mira(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether `x` is an object of class `mira`
}
\description{
Check for `mira` object
}
