% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mitml.result}
\alias{is.mitml.result}
\title{Check for `mitml.result` object}
\usage{
is.mitml.result(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether `x` is an object of class `mitml.result`
}
\description{
Check for `mitml.result` object
}
