% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.predictors.R
\name{lars.internal}
\alias{lars.internal}
\title{Filter out predictors by least angular regression}
\usage{
lars.internal(
  x,
  y,
  type = "lar",
  intercept = TRUE,
  lars.relax = 5,
  minimal.cp = 1,
  ...
)
}
\arguments{
\item{x}{
matrix of predictors
}

\item{y}{
response 
}

\item{type}{
One of  "lasso", "lar", "forward.stagewise" or "stepwise". The names can
be abbreviated to any unique substring. Default is "lasso".
}

\item{intercept}{
if TRUE, an intercept is included in the model (and not penalized),
otherwise no intercept is included. Default is TRUE.
}

\item{lars.relax}{numeric. The percentage of the minimum Cp value that is
added to the minimum Cp to relax the inclusion threshold. The default is 5.
Use 1-5 percent for a more stringent filter, and 5-10 percent for a moderate
filter.}

\item{minimal.cp}{numeric. The minimum "Cp" value to consider. "Cp" may
become negative for small samples. \code{minimal_cp} is the minimum "Cp" value
that is is used as to define the threshold for the filter. Higher value
select more predictors. The default is 1.}

\item{\dots}{Further arguments passed to \code{lars::lars}. Nothing passed down.}
}
\value{
A logical vector of length \code{ncol(x)} indicating which predictors
to keep.
}
\description{
Filter out predictors by least angular regression
}
