% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larspred.R
\name{larspred}
\alias{larspred}
\title{Quick selection of predictors by Least Angle Regression (LAR) algorithm}
\usage{
larspred(
  data,
  type = "lar",
  s = ncol(data) - 1,
  max.steps = ncol(data) - 1,
  ...
)
}
\arguments{
\item{data}{Matrix or data frame with incomplete data.}

\item{type}{A string specifying the type of LARS algorithm. Use
\code{'lar'} (default) or \code{'lasso'}. Can be abbreviated.}

\item{s}{See \code{\link{lars::predict.lars}} for details.}

\item{max.steps}{See \code{\link{lars::predict.lars}} for details.}

\item{...}{Passed down to the \code{\link{lars::lars}} and
\code{\link{lars::coef.lars}} functions.}
}
\value{
A square binary matrix of size \code{ncol(data)}.
}
\description{
Single imputation of incomplete data, followed by the LARS algorithm
for variables selection.
}
\details{
This function creates a predictor matrix using the variable selection
procedure described in Van Buuren et al.~(1999, p.~687--688). The function is
designed to aid in setting up a good imputation model for data with many
variables.

Basic workings: The procedure creates a single imputation of the data
by a random draw from the marginal, and then applies the LAR algorithm
twice for each incomplete variable: once with the variable as dependent,
and once with the missingness indicator of the variable as dependent. The
union of both LAR analyses is used to select the predictors for the
imputation model.

The LAR algorithm is a fast approximation of the LASSO algorithm, which
is a penalized regression method that shrinks coefficients to zero.
Predictors are selected based on the absolute correlation
between the target and the predictor. The procedure is fast and can
handle large datasets.

Problem: Imputation by random draws from the marginal distribution
weakens the correlation structure in the data. This may lead to
suboptimal predictor selection. The procedure is best used for
exploratory data analysis.

Potential extensions: The function can be extended to include other
imputation methods, to add support for categorical variables, and
to select the optimal number of variables by Mallow's Cp.
}
\note{
\code{larspred()} uses \code{\link[base]{data.matrix}} to convert
factors to numbers through their internal codes. For unordered factors
the resulting quantification may not make sense.
}
\examples{
# include all variables
larspred(nhanes)

# include two best variables
larspred(nhanes, s = 2)

# limit to three best variables
larspred(boys, s = 3)
}
\author{
Stef van Buuren, June 2024
}
