\name{leiden85}
\alias{leiden85}
\docType{data}
\title{Leiden 85+ Study}
\description{
Subset of data from the Leiden 85+ Study}
\usage{
data(leiden85)
}
\format{
  \code{leiden85} is a data frame with 956 rows and 336 columns.
}
\details{
The data set concerns of subset of 956 members of a very old (85+) cohort in Leiden.

Multiple imputation of this data set has been described in 
Boshuizen et al (1998), Van Buuren et al (1999) and 
Van Buuren (2012), chapter 7.

The data set is not available as part of \code{mice}.
}


\source{

Lagaay, A. M., van der Meij, J. C., Hijmans, W. (1992). 
Validation of medical history taking as part of a population based survey in subjects aged 85 and over.
\emph{Brit. Med. J.}, \emph{304}(6834), 1091-1092.

Izaks, G. J., van Houwelingen, H. C., Schreuder, G. M., Ligthart, G. J. (1997). 
The association between human leucocyte antigens (HLA) and mortality in community residents 
aged 85 and older. 
\emph{Journal of the American Geriatrics Society}, \emph{45}(1), 56-60.  

Boshuizen, H. C., Izaks, G. J., van Buuren, S., Ligthart, G. J. (1998). 
Blood pressure and mortality in elderly people aged 85 and older: Community based study. 
\emph{Brit. Med. J.}, \emph{316}(7147), 1780-1784. 
\url{http://www.stefvanbuuren.nl/publications/Blood pressure - BMJ 1998.pdf}

Van Buuren, S., Boshuizen, H.C., Knook, D.L. (1999) 
Multiple imputation of missing blood pressure covariates in survival analysis. 
\emph{Statistics in Medicine}, \bold{18}, 681--694. 
\url{http://www.stefvanbuuren.nl/publications/Multiple imputation - Stat Med 1999.pdf}

van Buuren, S. (2012). \emph{Flexible Imputation of Missing Data.} Boca Raton, FL: Chapman & Hall/CRC Press. 

}

\keyword{datasets}

