% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.mids}
\alias{lm.mids}
\title{Linear regression for `mids` object}
\usage{
lm.mids(formula, data, ...)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~
operator, and the terms, separated by + operators, on the right. See the
documentation of [lm()] and [formula()] for details.}

\item{data}{An object of type 'mids', which stands for 'multiply imputed data
set', typically created by a call to function `mice()`.}

\item{\dots}{Additional parameters passed to [lm()]}
}
\value{
An objects of class `mira`, which stands for 'multiply imputed
repeated analysis'.  This object contains `data$m` distinct
`lm.objects`, plus some descriptive information.
}
\description{
Applies `lm()` to multiply imputed data set
}
\details{
This function is included for backward compatibility with V1.0. The function
is superseded by [with.mids()].
}
\examples{
imp <- mice(nhanes)
fit <- lm.mids(bmi ~ hyp + chl, data = imp)
fit
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}
}
\seealso{
[lm()], [`mids()`][mids-class], [`mira()`][mira-class]
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000
}
\keyword{multivariate}
