% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2mids.r
\name{long2mids}
\alias{long2mids}
\title{Conversion of a imputed data set (long form) to a \code{mids} object}
\usage{
long2mids(x)
}
\arguments{
\item{x}{argument}
}
\value{
An object of class \code{mids}
}
\description{
Deprecated. Use \code{as.mids()} instead.
}
\author{
Stef van Buuren, 2012
}
\seealso{
\code{\link{as.mids}}
}
\keyword{internal}

