% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{make.dots}
\alias{make.dots}
\title{Creates a `dots` argument}
\usage{
make.dots(data, blocks = make.blocks(data))
}
\arguments{
\item{data}{A `data.frame` with the source data}

\item{blocks}{An optional specification for blocks of variables in
the rows. The default assigns each variable in its own block.}
}
\value{
A matrix
}
\description{
This helper function creates a valid `dots` object. The
`dots` object is an argument to the `mice` function.
The name `dots` is a contraction of blocks-dots.
Through `dots`, the user can specify any additional
arguments that are specifically passed down to the lowest level
imputation function.
}
\examples{
make.dots(nhanes, blocks = name.blocks(c("age", "hyp"), "xxx"))
}
\seealso{
[make.blocks()]
}
