% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{make.formulas}
\alias{make.formulas}
\title{Creates a \code{formulas} object}
\usage{
make.formulas(data, blocks = make.blocks(data), full = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} with the source data}

\item{blocks}{An optional specification for blocks of variables in 
the rows. The default assigns each variable in its own block.}

\item{full}{A logical indicating whether the full model (default) should be 
produced. The full model imputes each variable conditional 
on all other variables. Setting \code{full = FALSE} yields the empty 
model.}
}
\value{
A list of formula's. If \code{format == "character"} then it os
}
\description{
This helper function creates a valid \code{formulas} object. The 
\code{formulas} object is an argument to the \code{mice} function. 
It is a list of formula's that specifies the target variables and 
the predictors by means of the standard \code{~} operator.
}
\examples{
make.formulas(nhanes)
make.formulas(nhanes, blocks = make.blocks(nhanes, "collect"))
}
\seealso{
\code{\link{make.blocks}}, \code{\link{make.predictorMatrix}}
}
