% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{make.method}
\alias{make.method}
\title{Creates a `method` argument}
\usage{
make.method(
  data,
  where = make.where(data),
  blocks = make.blocks(data),
  defaultMethod = c("pmm", "logreg", "polyreg", "polr"),
  ynames = NULL
)
}
\arguments{
\item{data}{Data frame with \eqn{n} rows and \eqn{p} columns with
incomplete data.  Missing values are coded as `NA`.}

\item{where}{A data frame or matrix of logicals with \eqn{n} rows
and \eqn{p} columns, indicating the cells of `data` for
which imputations are generated.
The default `where = is.na(data)` specifies that all
missing data are imputed.
The `where` argument can overimpute cells
with observed data, or skip imputation of specific missing
cells. Be aware that the latter option could propagate
missing values to other variables. See details.
Note: Not all imputation methods may support the `where`
argument (e.g., `mice.impute.jomoImpute()` or
`mice.impute.panImpute()`).}

\item{blocks}{List of \eqn{q} character vectors that identifies the
variable names per block. The name of list elements
identify blocks. `mice()` will provide default names
(`"b1"`, `"b2"`, ...) for blocks containing multiple
variables. Variables within a block are imputed as a
block, e.g. by a multivariate imputation method, or
by an iterated version of the same univariate imputation
method. By default each variable is allocated to a
separate block, which is effectively fully conditional
specification (FCS) by univariate models
(variable-by-variable imputation).
All data variables are assigned to a block.
A variable can belong to only one block, so there are
at most \eqn{p} blocks.
See the `parcel` argument for an easier alternative to
the `blocks` argument.}

\item{defaultMethod}{A vector of length 4 containing the default imputation
methods for
1) numeric data (`"pmm"`)
2) factor data with 2 levels, (`"logreg"`)
3) factor data with > 2 unordered levels, (`"polyreg"`) and
4) factor data with > 2 ordered levels (`"polr"`).
The `defaultMethod` can be used to alter to default mapping
of variable type to imputation method.}

\item{ynames}{vector of names of variables to be imputed}
}
\value{
Vector of `length(blocks)` element with method names
}
\description{
This helper function creates a valid `method` vector. The
`method` vector is an argument to the `mice` function that
specifies the method for each block.
}
\examples{
make.method(nhanes2)
}
\seealso{
[mice()]
}
