% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{make.nest}
\alias{make.nest}
\title{Creates a \code{nest} argument}
\usage{
make.nest(x, partition = c("scatter", "collect", "void"), prefix = "A")
}
\arguments{
\item{x}{A \code{data.frame}, an unnamed character vector, a named
character vector or a \code{list}.}

\item{partition}{Only relevant if \code{x} is a \code{data.frame}. Value
\code{"scatter"} (default) will assign each variable to a separate
nest. Value \code{"collect"} assigns all variables to one nest,
whereas \code{"void"} does not assign any variable to a nest.}

\item{prefix}{A character vector of length 1 with the prefix to
be using for naming any unnamed blocks with two or more variables.}
}
\value{
A character vector of length \code{ncol(data)} that specifies
the nest name per variable
}
\description{
This helper function generates a character vector for the
\code{nest} argument in the \code{\link[=mice]{mice()}} function.
}
\details{
Choices \code{"scatter"} and \code{"collect"} represent to two
extreme scenarios for assigning variables to imputation nests.
Use \code{"scatter"} to create an imputation model based on
\emph{fully conditionally specification} (FCS). Use \code{"collect"} to
gather all variables to be imputed by a \emph{joint model} (JM).

Any variable not listed in the result will not be imputed.
Specification \code{"void"} represents the extreme scenario where
nothing is imputed.

Unlike blocks, a variable cannot be allocated to multiple nests.
}
\examples{

# default nest creation (scatter)
make.nest(nhanes)

# make nest from variable names
make.nest(c("age", "sex", "edu"))

# put hgt, wgt and bmi into one nest, automatic naming
make.nest(list("age", "sex", c("hgt", "wgt", "bmi")))

# same, but with custom nest names
make.nest(list("age", "sex", anthro = c("hgt", "wgt", "bmi")))

# all variables into one nest
make.nest(nhanes, partition = "collect", prefix = "myblock")
}
