% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make.newdata}
\alias{make.newdata}
\title{Construct a new data.frame from a trained model}
\usage{
make.newdata(models, n = 1L, fill = NA, vars = NULL)
}
\arguments{
\item{models}{A list of trained models from \code{mice()}.}

\item{n}{Number of rows to generate. Default is 1.}

\item{fill}{Value to fill the data with. Default is \code{NA}.}

\item{vars}{Optional character vector specifying the variable order, or for
a selecting a subset of variable. The default returns all variables in
the order as they appear in \code{models}. Beware that taking a subset could lead
to an error if a predictor is undefined in the generated data.}
}
\value{
A \code{data.frame} of \code{n} rows, where each column matches the type and levels
expected from the corresponding model.
}
\description{
This function generates a data.frame with the correct structure (names, classes,
levels) based on a set of trained models, as produced by \code{mice()} with \code{tasks = "train"}.
}
\examples{
trained <- mice(boys, tasks = "train", print = FALSE)
newdata <- make.newdata(trained$models, n = 5, vars = names(boys))
str(newdata)
}
