% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{make.tasks}
\alias{make.tasks}
\title{Creates a \code{tasks} argument}
\usage{
make.tasks(data, tasks = "impute", blocks = make.blocks(data))
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as \code{NA}.}

\item{tasks}{A character vector specifying the task to perform for
each imputation block. The available options are:
\describe{
\item{"impute"}{Estimate parameters, generates imputations and store
the original data plus imputations, but not the imputation model
(classic MICE behavior).}
\item{"train" }{Estimate parameters, generate imputations and
store the original data, the imputations and the imputation model.}
\item{"fill"}{Apply a previously trained imputation model to fill
imputations, without re-estimating parameters.}
}
This argument can be specified as a named vector, where names correspond
to variables and values specify the task for each variable. If a
single value is provided, it applies to the variables in all blocks. The
length of the vector must match the number of variables present in the
blocks. The default is \code{"impute"}.}

\item{blocks}{List of vectors with variable names per block. List elements
may be named to identify blocks. Variables within a block are
imputed by a multivariate imputation method
(see \code{method} argument). By default each variable is placed
into its own block, which is effectively
fully conditional specification (FCS) by univariate models
(variable-by-variable imputation). Only variables whose names appear in
\code{blocks} are imputed. The relevant columns in the \code{where}
matrix are set to \code{FALSE} of variables that are not block members.
A variable may appear in multiple blocks. In that case, it is
effectively re-imputed each time that it is visited.}
}
\value{
Character vector of \code{ncol(data)} elements
}
\description{
This helper function creates a valid \code{tasks} vector. The
\code{tasks} vector is an argument to the \code{mice} function that
specifies the task for each column in the data.
}
\examples{
make.tasks(nhanes2)
}
\seealso{
\code{\link{mice}}
}
