% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{make.where}
\alias{make.where}
\title{Creates a `where` argument}
\usage{
make.where(data, keyword = c("missing", "all", "none", "observed"))
}
\arguments{
\item{data}{A `data.frame` with the source data}

\item{keyword}{An optional keyword, one of `"missing"` (missing
values are imputed), `"observed"` (observed values are imputed),
`"all"` and `"none"`. The default
is `keyword = "missing"`}
}
\value{
A matrix with logical
}
\description{
This helper function creates a valid `where` matrix. The
`where` matrix is an argument to the `mice` function.
It has the same size as `data` and specifies which values
are to be imputed (`TRUE`) or nor (`FALSE`).
}
\examples{
head(make.where(nhanes), 3)

# create & analyse synthetic data
where <- make.where(nhanes2, "all")
imp <- mice(nhanes2,
  m = 10, where = where,
  print = FALSE, seed = 123
)
fit <- with(imp, lm(chl ~ bmi + age + hyp))
summary(pool.syn(fit))
}
\seealso{
[make.blocks()], [make.predictorMatrix()]
}
