% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-package.R
\docType{package}
\name{mice-package}
\alias{mice-package}
\title{Multivariate Imputation by Chained Equations (MICE)}
\description{
The \code{mice} package implements a method to deal with missing data. The package
creates multiple imputations (replacement values) for multivariate missing data.
The method is based on Fully Conditional Specification, where each incomplete
variable is imputed by a separate model. The MICE algorithm can impute mixes of
continuous, binary, unordered categorical, and ordered categorical data.
Additionally, MICE can impute continuous two-level data and maintain consistency
between imputations using passive imputation. Many diagnostic plots are implemented
to inspect the quality of the imputations.

The \code{mice} package contains functions to:
\itemize{
\item Inspect the missing data pattern
\item Impute the missing data \emph{m} times, resulting in \emph{m} completed data sets
\item Diagnose the quality of the imputed values
\item Analyze each completed data set
\item Pool the results of the repeated analyses
\item Store and export the imputed data in various formats
\item Generate simulated incomplete data
\item Incorporate custom imputation methods
}
}
\section{Functions}{


The main functions are:\tabular{ll}{
   Function \tab Description \cr
   \code{mice()} \tab Impute the missing data \emph{m} times \cr
   \code{with()} \tab Analyze completed data sets \cr
   \code{pool()} \tab Combine parameter estimates \cr
   \code{complete()} \tab Export imputed data \cr
   \code{ampute()} \tab Generate missing data \cr
}
}

\section{Vignettes}{


A detailed series of six online vignettes walk you through solving realistic
inference problems with \code{mice}. We suggest going through these vignettes in the
following order:
\enumerate{
\item \href{https://www.gerkovink.com/miceVignettes/Ad_hoc_and_mice/Ad_hoc_methods.html}{Ad hoc methods and the MICE algorithm}
\item \href{https://www.gerkovink.com/miceVignettes/Convergence_pooling/Convergence_and_pooling.html}{Convergence and pooling}
\item \href{https://www.gerkovink.com/miceVignettes/Missingness_inspection/Missingness_inspection.html}{Inspecting how the observed data and missingness are related}
\item \href{https://www.gerkovink.com/miceVignettes/Passive_Post_processing/Passive_imputation_post_processing.html}{Passive imputation and post-processing}
\item \href{https://www.gerkovink.com/miceVignettes/Multi_level/Multi_level_data.html}{Imputing multilevel data}
\item \href{https://www.gerkovink.com/miceVignettes/Sensitivity_analysis/Sensitivity_analysis.html}{Sensitivity analysis with \code{mice}}
}

Van Buuren, S. (2018). \emph{Flexible Imputation of Missing Data. Second Edition.}
Boca Raton, FL: Chapman & Hall/CRC Press.
The book \href{https://stefvanbuuren.name/fimd/}{Flexible Imputation of Missing Data} contains
\href{https://github.com/stefvanbuuren/fimdbook/tree/master/R}{example code}.
}

\section{Methodology}{


The \code{mice} software was published in the \emph{Journal of Statistical Software}
(Van Buuren and Groothuis-Oudshoorn, 2011).
\href{https://doi.org/10.18637/jss.v045.i03}{doi:10.18637/jss.v045.i03}.
The first application of the method concerned missing blood pressure data
(Van Buuren et al., 1999). The term \emph{Fully Conditional Specification} was
introduced in 2006 to describe a general class of methods that specify
imputation models for multivariate data as a set of conditional distributions
(Van Buuren et al., 2006). Further details on mixes of variables and applications
can be found in the book \href{https://stefvanbuuren.name/fimd/}{Flexible Imputation of Missing Data}.
Chapman & Hall/CRC. Boca Raton, FL.
}

\section{Enhanced linear algebra}{


Updating the BLAS can improve the speed of R, sometimes considerably. The details
depend on the operating system. See the discussion in the
"R Installation and Administration" guide for further information.
}

\references{
van Buuren, S., Boshuizen, H.C., Knook, D.L. (1999).
Multiple imputation of missing blood pressure covariates in survival analysis.
\emph{Statistics in Medicine}, \strong{18}, 681–694.

van Buuren, S., Brand, J.P.L., Groothuis-Oudshoorn, C.G.M., Rubin, D.B. (2006).
Fully conditional specification in multivariate imputation.
\emph{Journal of Statistical Computation and Simulation}, \strong{76}(12), 1049–1064.

van Buuren, S., Groothuis-Oudshoorn, K. (2011).
\emph{mice: Multivariate Imputation by Chained Equations in R}.
\emph{Journal of Statistical Software}, \strong{45}(3), 1–67.
\href{https://doi.org/10.18637/jss.v045.i03}{doi:10.18637/jss.v045.i03}

Van Buuren, S. (2018).
\href{https://stefvanbuuren.name/fimd/}{Flexible Imputation of Missing Data}.
Chapman & Hall/CRC. Boca Raton, FL.
}
\seealso{
\itemize{
\item \href{mice}{\code{mice()}}
\item \href{with.mids}{\code{with.mids()}}
\item \href{pool}{\code{pool()}}
\item \href{complete}{\code{complete()}}
\item \href{ampute}{\code{ampute()}}
}
}
\author{
\strong{Maintainer}: Stef van Buuren \email{stef.vanbuuren@tno.nl}

Authors:
\itemize{
  \item Karin Groothuis-Oudshoorn \email{c.g.m.oudshoorn@utwente.nl}
}

Other contributors:
\itemize{
  \item Gerko Vink \email{g.vink@uu.nl} [contributor]
  \item Rianne Schouten \email{R.M.Schouten@uu.nl} [contributor]
  \item Alexander Robitzsch \email{robitzsch@ipn.uni-kiel.de} [contributor]
  \item Patrick Rockenschaub \email{rockenschaub.patrick@gmail.com} [contributor]
  \item Lisa Doove \email{lisa.doove@ppw.kuleuven.be} [contributor]
  \item Shahab Jolani \email{s.jolani@maastrichtuniversity.nl} [contributor]
  \item Margarita Moreno-Betancur \email{margarita.moreno@mcri.edu.au} [contributor]
  \item Ian White \email{ian.white@ucl.ac.uk} [contributor]
  \item Philipp Gaffert \email{philipp.gaffert@gfk.com} [contributor]
  \item Florian Meinfelder \email{florian.meinfelder@uni-bamberg.de} [contributor]
  \item Bernie Gray \email{bfgray3@gmail.com} [contributor]
  \item Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca} [contributor]
  \item Mingyang Cai \email{m.cai@uu.nl} [contributor]
  \item Thom Volker \email{t.b.volker@uu.nl} [contributor]
  \item Edoardo Costantini \email{e.costantini@tilburguniversity.edu} [contributor]
  \item Caspar van Lissa \email{c.j.vanlissa@uu.nl} [contributor]
  \item Hanne Oberman \email{h.i.oberman@uu.nl} [contributor]
  \item Stephen Wade \email{stephematician@gmail.com} [contributor]
}

}
\keyword{internal}
