% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.heckman.R
\name{mice.impute.2l.heckman}
\alias{mice.impute.2l.heckman}
\alias{2l.heckman}
\title{Imputation based on Heckman model for multilevel data.}
\usage{
mice.impute.2l.heckman(
  y,
  ry,
  x,
  wy = NULL,
  type,
  pmm = FALSE,
  meta_method = "reml",
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{type}{type of the variable in the prediction model {0: No predictor,
1: Predictor in both the outcome and selection,-2: Cluster id (study id),
-3: Predictor only in the selection model, -4: Predictor only in the outcome 
model}}

\item{pmm}{predictive mean matching can be applied only for for missing continuous variables: "TRUE","FALSE"}

\item{meta_method}{meta_analysis estimation method for random effects :
"ml" (maximum likelihood), "reml" (restricted maximum likelihood) or "mm"
method of moments.}

\item{...}{Other named arguments. Not used.}
}
\value{
Vector with imputed data, of type binary or continuous
}
\description{
Imputes outcome and predictor variables that follow an MNAR mechanism
according to Heckman's model and come from a multilevel database such as
individual participant data.
}
\details{
Imputes systematically and sporadically missing binary and continuous
 univariate variables that follow a MNAR mechanism according to the Heckman 
 selection model and come from a clustered dataset. 
 The imputation method uses a two-stage approach in which the Heckman model 
 parameters at the cluster level are estimated using the copula method.
}
\note{
Missing binary variables should be included as two-level factor type variables. 
When the cluster variable is not defined in the predictor matrix as "-2", the imputation
method is based on a simple Heckman model, i.e. without taking into account 
the hierarchical structure.
In case the Heckman model cannot be estimated at the study level, the imputation 
method will be based on the simple Heckman model.
Added:
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\author{
Julius Center Methods Group UMC, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
