% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.rbart.R
\name{mice.impute.2l.rbart}
\alias{mice.impute.2l.rbart}
\title{A short description of your function}
\usage{
mice.impute.2l.rbart(
  y,
  ry,
  x,
  wy = NULL,
  type,
  use.matcher = FALSE,
  donors = 5L,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{type}{Vector of length \code{ncol(x)} identifying random and class
variables.  Random variables are identified by a '2'. The class variable
(only one is allowed) is coded as '-2'. Fixed effects are indicated by
a '1'.}

\item{use.matcher}{Logical. Set \code{use.matcher = TRUE} to specify
the C function \code{matcher()}, the now deprecated matching function that
was default in versions
\code{2.22} (June 2014) to \code{3.11.7} (Oct 2020). Since version \code{3.12.0}
\code{mice()} uses the much faster \code{matchindex} C function. Use
the deprecated \code{matcher} function only for exact reproduction.}

\item{donors}{The size of the donor pool among which a draw is made.
The default is \code{donors = 5L}. Setting \code{donors = 1L} always selects
the closest match, but is not recommended. Values between 3L and 10L
provide the best results in most cases (Morris et al, 2015).}

\item{...}{Other named arguments.}

\item{intercept}{Logical determining whether the intercept is automatically
added.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}.
}
\description{
A short description of your function
}
