% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.sys.R
\name{mice.impute.2l.sys}
\alias{mice.impute.2l.sys}
\title{Imputation for systematic missing data by a two-level normal model}
\usage{
mice.impute.2l.sys(y, ry, x, type, wy = NULL, intercept = TRUE, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the 
the subset \code{y[ry]} of elements in \code{y} to which the imputation 
model is fitted. The \code{ry} generally distinguishes the observed 
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for 
\code{y}. Matrix \code{x} may have no missing values.}

\item{type}{Vector of length \code{ncol(x)} identifying random and class
variables.  Random variables are identified by a '2'. The class variable
(only one is allowed) is coded as '-2'. Fixed effects are indicated by 
a '1'.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value 
indicates locations in \code{y} for which imputations are created.}

\item{intercept}{Logical determining whether the intercept is automatically
added.}

\item{...}{Other named arguments.}
}
\value{
A vector of length \code{nmis} with imputations.
}
\description{
Imputes univariate systematically missing data using a two-level normal model
}
\details{
Data are missing systematically if they have not been measured, e.g., in the 
case where we combine data from different souces.
}
\references{
Jolani S, Debray TPA, Koffijberg H, van Buuren S, Moons KGM (2015).
Imputation of systematically missing predictors in an individual 
participant data meta-analysis: a generalized approach using MICE. 
\emph{Statistics in Medicine}, 34:1841-1863.

Van Buuren, S. (2011) Multiple imputation of multilevel data. In Hox, J.J.
and and Roberts, J.K. (Eds.), \emph{The Handbook of Advanced Multilevel
Analysis}, Chapter 10, pp. 173--196. Milton Park, UK: Routledge.
}
\seealso{
Other univariate \code{2l} functions: \code{\link{mice.impute.2l.norm}}
}
\author{
Shahab Jolani, 2017
}
\keyword{datagen}
