% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2lonly.mean.R
\name{mice.impute.2lonly.mean}
\alias{mice.impute.2lonly.mean}
\alias{2lonly.mean}
\title{Imputation of most likely value within the class}
\usage{
mice.impute.2lonly.mean(y, ry, x, type, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{type}{Vector of length `ncol(x)` identifying random and class
variables.  The class variable (only one is allowed) is coded as `-2`.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Method `2lonly.mean` replicates the most likely value within
a class of a second-level variable. It works for numeric and
factor data. The function is primarily useful as a quick fixup for
data in which the second-level variable is inconsistent.
}
\details{
Observed values in `y` are averaged within the class, and
replicated to the missing `y` within that class.
This function is primarily useful for repairing incomplete data
that are constant within the class, but vary over classes.

For numeric variables, `mice.impute.2lonly.mean()` imputes the
class mean of `y`. If `y` is a second-level variable, then
conventionally all observed `y` will be identical within the
class, and the function just provides a quick fix for any
missing `y` by filling in the class mean.

For factor variables, `mice.impute.2lonly.mean()` imputes the
most frequently occuring category within the class.

If there are no observed `y` in the class, all entries of the
class are set to `NA`. Note that this may produce problems
later on in `mice` if imputation routines are called that
expects predictor data to be complete. Methods designed for
imputing this type of second-level variables include
[mice.impute.2lonly.norm()] and
[mice.impute.2lonly.pmm()].
}
\references{
Van Buuren, S. (2018).
[*Flexible Imputation of Missing Data. Second Edition.*](https://stefvanbuuren.name/fimd/sec-level2pred.html)
Boca Raton, FL.: Chapman & Hall/CRC Press.
}
\seealso{
Other univariate-2lonly: 
\code{\link{mice.impute.2lonly.norm}()},
\code{\link{mice.impute.2lonly.pmm}()}
}
\author{
Gerko Vink, Stef van Buuren, 2019
}
\concept{univariate-2lonly}
\keyword{datagen}
