% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.gspcr.pmm.R
\name{mice.impute.gspcr.pmm}
\alias{mice.impute.gspcr.pmm}
\alias{gspcr.pmm}
\title{Imputation by predictive mean matching based on generalized supervised principal component regression}
\usage{
mice.impute.gspcr.pmm(
  y,
  ry,
  x,
  wy = NULL,
  thrs = "PR2",
  fit_measure = "BIC",
  nthrs = 10,
  npcs_range = 1:3,
  K = 1,
  donors = 5L,
  exclude = -99999999,
  use.matcher = FALSE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{thrs}{character vector of length 1 storing the type of threshold to be used (see below for available options)}

\item{fit_measure}{character vector of length 1 indicating the type of fit measure to be used in the cross-validation procedure}

\item{nthrs}{numeric vector of length 1 storing the number of threshold values to be used}

\item{npcs_range}{numeric vector defining the numbers of principal components to be used}

\item{K}{numeric vector of length 1 storing the number of folds for the K-fold cross-validation procedure}

\item{donors}{the size of the donor pool among which a draw is made (default is \code{5L}).}

\item{exclude}{Value or vector of values to exclude from the imputation donor pool in \code{y}}

\item{use.matcher}{Logical. Set \code{use.matcher = TRUE} to specify
the C function \code{matcher()}, the now deprecated matching function that
was default in versions
\code{2.22} (June 2014) to \code{3.11.7} (Oct 2020). Since version \code{3.12.0}
\code{mice()} uses the much faster \code{matchindex} C function. Use
the deprecated \code{matcher} function only for exact reproduction.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, the same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data using predictive mean matching and generalized supervised principal component regression.
}
\details{
For details on the gspcr framework, see [mice::mice.impute.gspcr.norm()].
For details on how the pmm framework see [mice::mice.impute.pmm()].
Note that \code{gspcr.pmm} uses match type 2, based on the bootstrap pmm presented in Heitjan and Little (1991).
}
\references{
Bair, E., Hastie, T., Paul, D., & Tibshirani, R. (2006). Prediction by
supervised principal components. Journal of the American Statistical
Association, 101(473), 119-137.

Heitjan, D. F., & Little, R. J. (1991). Multiple imputation for the fatal accident reporting system. Journal of the Royal Statistical Society Series C: Applied Statistics, 40(1), 13-29.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.gspcr.logreg}()},
\code{\link{mice.impute.gspcr.norm}()},
\code{\link{mice.impute.gspcr.polr}()},
\code{\link{mice.impute.gspcr.polyreg}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{imputation}
