% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.gspcr.polyreg.R
\name{mice.impute.gspcr.polyreg}
\alias{mice.impute.gspcr.polyreg}
\alias{gspcr.polyreg}
\title{Imputation by generalized supervised principal component polytomous regression.}
\usage{
mice.impute.gspcr.polyreg(
  y,
  ry,
  x,
  wy = NULL,
  thrs = "PR2",
  fit_measure = "BIC",
  nthrs = 10,
  npcs_range = 1:5,
  K = 1,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{thrs}{character vector of length 1 storing the type of threshold to be used (see below for available options)}

\item{fit_measure}{character vector of length 1 indicating the type of fit measure to be used in the cross-validation procedure}

\item{nthrs}{numeric vector of length 1 storing the number of threshold values to be used}

\item{npcs_range}{numeric vector defining the numbers of principal components to be used}

\item{K}{numeric vector of length 1 storing the number of folds for the K-fold cross-validation procedure}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, the same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing values on an unorded categorical (nominal) variable using generalized supervised principal component proportional odds regression.
}
\details{
See [mice.impute.gspcr.norm].
}
\references{
Bair, E., Hastie, T., Paul, D., & Tibshirani, R. (2006). Prediction by
supervised principal components. Journal of the American Statistical
Association, 101(473), 119-137.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.gspcr.logreg}()},
\code{\link{mice.impute.gspcr.norm}()},
\code{\link{mice.impute.gspcr.pmm}()},
\code{\link{mice.impute.gspcr.polr}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{imputation}
