% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.heckman.ipd.R
\name{mice.impute.heckman.ipd}
\alias{mice.impute.heckman.ipd}
\alias{heckman.ipd}
\title{Imputation based on Heckman model for Individual Patient Data.}
\usage{
mice.impute.heckman.ipd(
  y,
  ry,
  x,
  wy = NULL,
  type,
  family,
  full = FALSE,
  pmm = FALSE,
  meta_method,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{type}{type of the variable in the prediction model {0: No predictor,
1: Predictor in both the outcome and selection,-2: Cluster id (study id),
-3: Predictor only in the selection model, -4: Predictor only in the outcome 
model}}

\item{family}{family of the missing variable {"normal","probit"}}

\item{full}{If full=TRUE all parameters are fully correlated otherwise the
correlation is given within each of the following parameter sets: selection
coefficient parameters, outcome coefficient parameters, rho and sigma}

\item{pmm}{for gaussian variables if applies predictive mean matching to
imputed result: "TRUE","FALSE"}

\item{meta_method}{meta_analysis estimation method for random effects :
"ml" (maximum likelihood), "reml" (restricted maximum likelihood) or "mm"
method of moments.}

\item{...}{Other named arguments. Not used.}
}
\value{
Vector with imputed data, of type binary or continuous
}
\description{
Imputes outcome and predictor variables that follow an MNAR mechanism
according to Heckman's model and come from a clustered database such as
individual participant data.
}
\details{
Imputation of binary and continuous variables by Heckman model.
This function uses information at the marginal and cluster level to impute
both outcome and predictor variables that are sporadically or systematically
missing in a cluster.
Initially, the function estimates the parameters of the Heckman model for
each cluster in the database using the copula method, and then estimates the
parameters at the marginal level with the help of a random intercept
meta-analysis model.
The individual-level parameters are drawn and missing values are imputed
using Heckman's conditional expectation.
}
\note{
In case the Heckman model cannot be estimated for a given cluster, missing
values in this cluster will be imputed without using the selection bias
correction, i.e. a MAR mechanism is assumed for this cluster.
Added:
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\author{
Julius Center Methods Group UMC, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
