% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.logreg.R
\name{mice.impute.logreg.boot}
\alias{mice.impute.logreg.boot}
\title{Imputation by logistic regression using the bootstrap}
\usage{
mice.impute.logreg.boot(y, ry, x, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Imputes univariate missing data using logistic regression
by a bootstrapped logistic regression model.
The bootstrap method draws a simple bootstrap sample with replacement
from the observed data `y[ry]` and `x[ry, ]`.
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}

Van Buuren, S. (2018).
[*Flexible Imputation of Missing Data. Second Edition.*](https://stefvanbuuren.name/fimd/sec-categorical.html)
Chapman & Hall/CRC. Boca Raton, FL.
}
\seealso{
[mice()], [glm()], [glm.fit()]

Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000, 2011
}
\concept{univariate imputation functions}
\keyword{datagen}
