% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.mean.R
\name{mice.impute.mean}
\alias{mice.impute.mean}
\title{Imputation by the mean}
\usage{
mice.impute.mean(y, ry, x = NULL, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Imputes the arithmetic mean of the observed data
}
\section{Warning}{
 Imputing the mean of a variable is almost never
appropriate.  See Little and Rubin (2002, p. 61-62) or
Van Buuren (2012, p. 10-11)
}

\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}

Little, R.J.A. and Rubin, D.B. (2002). Statistical Analysis with Missing
Data.  New York: John Wiley and Sons.

Van Buuren, S. (2018).
[*Flexible Imputation of Missing Data. Second Edition.*](https://stefvanbuuren.name/fimd/sec-simplesolutions.html#sec:meanimp)
Chapman & Hall/CRC. Boca Raton, FL.
}
\seealso{
[mice()], [mean()]

Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()}
}
\concept{univariate imputation functions}
\keyword{datagen}
