% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pcr.R
\name{mice.impute.pcr}
\alias{mice.impute.pcr}
\alias{pcr}
\title{Imputation by principal component regression}
\usage{
mice.impute.pcr(y, ry, x, wy = NULL, npcs = 1L, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{npcs}{The number of principal components to extract in the PC regression.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data by extracting PCs from all potential auxiliary variables.
}
\details{
Imputation of \code{y} by principal component regression (Jolliffe, 1986, p. 168).
The method consists of the following steps:
\enumerate{
\item For a given \code{y} variable under imputation, draw a bootstrap version y*
with replacement from the observed cases \code{y[ry]}, and stores in x* the
corresponding values from \code{x[ry, ]}.
\item Fit a PC regression with y* as the outcome, and x* as predictors using 
\code{npcs} components.
\item Calculate the estimated residual standard error \code{sigma} based on the residuals 
obtained from the PC regression and \code{npcs - 1} degrees of freedom.
\item Obtain predicted values for \code{y} based on the fitted PC regression 
and the new data \code{x[wy, ]} 
\item Obtain imputations by adding noise scaled by \code{sigma} to these
predictions.
}
}
\references{
Jolliffe, I. (1986) Principal Component Analysis. Springer.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pcovr}()},
\code{\link{mice.impute.pls}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()},
\code{\link{mice.impute.spcr}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
