% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pls.R
\name{mice.impute.pls}
\alias{mice.impute.pls}
\alias{pls}
\title{Imputation by partial least square regression}
\usage{
mice.impute.pls(y, ry, x, wy = NULL, nlvs = 1L, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{nlvs}{The number of latent variables to consider in the PLS regression.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data by using partial least square regression.
}
\details{
Imputation of \code{y} by PLS. The procedure is as follows:
\enumerate{
\item A bootstrap sample of the observed part of the data is taken
\item PLSR is estimated on this sample for a given number of latent variables (nlvs).
\item Based on this PLSR model, predictions on the new data x[!ry, ]
\item Imputations are then obtained by adding noise to these predictions.
}

Noise is added by sampling from a standard normal distribution rescaled based 
on the residuals standard error of the PLSR.
The residual degrees of freedom (df) are computed according to a naive approach 
(n - nlvs - 1).
Future developments of this function could include one of the more elaborate 
approaches to estimating df described in Kramer and Sugiyama (2011).
}
\references{
Krämer, N., & Sugiyama, M. (2011). The degrees of freedom of partial least 
squares regression. Journal of the American Statistical Association, 106(494), 697-705.

Wehrens, R., & Mevik, B. H. (2007). The pls package: principal component and 
partial least squares regression in R.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pcovr}()},
\code{\link{mice.impute.pcr}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()},
\code{\link{mice.impute.spcr}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
