% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.ri.R
\name{mice.impute.ri}
\alias{mice.impute.ri}
\alias{ri}
\title{Imputation by the random indicator method for nonignorable data}
\usage{
mice.impute.ri(y, ry, x, wy = NULL, ri.maxit = 10, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{ri.maxit}{Number of inner iterations}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Imputes nonignorable missing data by the random indicator method.
}
\details{
The random indicator method estimates an offset between the
distribution of the observed and missing data using an algorithm
that iterates over the response and imputation models.

This routine assumes that the response model and imputation model
have same predictors.

For an MNAR alternative see also [mice.impute.mnar.logreg()].
}
\references{
Jolani, S. (2012).
*Dual Imputation Strategies for Analyzing Incomplete Data*.
Dissertation. University of Utrecht, Dec 7 2012.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()}
}
\author{
Shahab Jolani (University of Utrecht)
}
\concept{univariate imputation functions}
\keyword{datagen}
