% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.sample.r
\name{mice.impute.sample}
\alias{mice.impute.sample}
\title{Imputation by simple random sampling}
\usage{
mice.impute.sample(y, ry, x = NULL, wy = NULL, ...)
}
\arguments{
\item{y}{Incomplete data vector}

\item{ry}{Vector of missing data pattern of \code{y}}

\item{x}{Matrix with \code{length(y)} rows with complete covariates. Not used.}

\item{wy}{Logical vector of length \code{length(y)} elements indicating 
observation for which imputations should be created}

\item{...}{Other named arguments. Not used.}
}
\value{
A vector of length \code{sum(wy)} with imputations
}
\description{
Imputes a random sample from the observed \code{y} data
}
\details{
This function takes a simple random sample from the observed values in
\code{y}, and returns these as imputations.
}
\references{
van Buuren S and Groothuis-Oudshoorn K (2011). \code{mice}:
Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of
Statistical Software}, \bold{45}(3), 1-67.
\url{http://www.jstatsoft.org/v45/i03/}
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000, 2017
}
\keyword{datagen}
