% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipo.R
\name{mipo}
\alias{mipo}
\alias{summary.mipo}
\alias{print.mipo}
\alias{print.mipo.summary}
\alias{process_mipo}
\title{`mipo`: Multiple imputation pooled object}
\usage{
mipo(mira.obj, ...)

\method{summary}{mipo}(
  object,
  type = c("tests", "all"),
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE,
  ...
)

\method{print}{mipo}(x, ...)

\method{print}{mipo.summary}(x, ...)

process_mipo(z, x, conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE)
}
\arguments{
\item{mira.obj}{An object of class `mira`}

\item{\dots}{Arguments passed down}

\item{object}{An object of class `mipo`}

\item{conf.int}{Logical indicating whether to include
a confidence interval.}

\item{conf.level}{Confidence level of the interval, used only if
`conf.int = TRUE`. Number between 0 and 1.}

\item{exponentiate}{Flag indicating whether to exponentiate the
coefficient estimates and confidence intervals (typical for
logistic regression).}

\item{x}{An object of class `mipo`}

\item{z}{Data frame with a tidied version of a coefficient matrix}
}
\value{
The `summary` method returns a data frame with summary statistics of the pooled analysis.
}
\description{
The `mipo` object contains the results of the pooling step.
The function [pool()] generates an object of class `mipo`.
}
\details{
An object class `mipo` is a `list` with
elements: `call`, `m`, `pooled` and `glanced`.

The `pooled` elements is a data frame with columns:
\tabular{ll}{
`estimate`\tab Pooled complete data estimate\cr
`ubar`    \tab Within-imputation variance of `estimate`\cr
`b`       \tab Between-imputation variance of `estimate`\cr
`t`       \tab Total variance, of `estimate`\cr
`dfcom`   \tab Degrees of freedom in complete data\cr
`df`      \tab Degrees of freedom of $t$-statistic\cr
`riv`     \tab Relative increase in variance\cr
`lambda`  \tab Proportion attributable to the missingness\cr
`fmi`     \tab Fraction of missing information\cr
}
The names of the terms are stored as `row.names(pooled)`.

The `glanced` elements is a `data.frame` with `m` rows.
The precise composition depends on the class of the complete-data analysis.
At least field `nobs` is expected to be present.

The `process_mipo` is a helper function to process a
tidied mipo object, and is normally not called directly.
It adds a confidence interval, and optionally exponentiates, the result.
}
\references{
van Buuren S and Groothuis-Oudshoorn K (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}
}
\seealso{
[pool()],
[`mids()`][mids-class], [`mira()`][mira-class]
}
\keyword{classes}
\keyword{internal}
