% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mira.R
\docType{class}
\name{mira-class}
\alias{mira-class}
\alias{mira}
\title{Multiply imputed repeated analyses (`mira`)}
\description{
The `mira` object is generated by the `with.mids()` function.
The `as.mira()`
function takes the results of repeated complete-data analysis stored as a
list, and turns it into a `mira` object that can be pooled.
}
\details{
In versions prior to `mice 3.0` pooling required only that
`coef()` and `vcov()` methods were available for fitted
objects. *This feature is no longer supported*. The reason is that `vcov()`
methods are inconsistent across packages, leading to buggy behaviour
of the `pool()` function. Since `mice 3.0+`, the `broom`
package takes care of filtering out the relevant parts of the
complete-data analysis. It may happen that you'll see the messages
like `No method for tidying an S3 object of class ...` or
`Error: No glance method for objects of class ...`. The royal
way to solve this problem is to write your own `glance()` and `tidy()`
methods and add these to `broom` according to the specifications
given in <https://broom.tidymodels.org>.

#'The `mira` class of objects has methods for the
following generic functions: `print`, `summary`.

Many of the functions of the `mice` package do not use the
S4 class definitions, and instead rely on the S3 list equivalent
`oldClass(obj) <- "mira"`.
}
\section{Slots}{

 \describe{
 #'    \item{`.Data`:}{Object of class `"list"` containing the
   following slots:}
   \item{`call`:}{The call that created the object.}
   \item{`call1`:}{The call that created the `mids` object that was used
in `call`.}
   \item{`nmis`:}{An array containing the number of missing observations per
column.}
   \item{`analyses`:}{A list of `m` components containing the individual
fit objects from each of the `m` complete data analyses.}
   }
}

\references{
van Buuren S and Groothuis-Oudshoorn K (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}
}
\seealso{
[with.mids()], [`mids()`][mids-class], [mipo()]
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000
}
\keyword{classes}
