\name{mira}
\alias{mira}
\alias{print.mira}
\alias{summary.mira}
\title{Multiply Imputed Repeated Analysis}
\usage{
print.mira(x,...)
summary.mira(object, correlation, ...)
}
\arguments{
    \item{x, object}{An object containing the m fit objects of a complete data analysis, 
        plus some additional information.}
    \item{correlation}{}
    \item{...}{not used}
}
\description{
  The "mira" object is generated by the lm.mids and glm.mids functions.
  The "mira" class of objects has methods for the following generic functions:
  print, summary.}

\value{
  \item{call}{  The call that created the object.}
  \item{call1}{ The call that created the mids object that was used in 'call'.}
  \item{nmis}{  An array containing the number of missing observations per column.}
  \item{analyses}{ A list of m components containing the individual fit objects from each of the m complete data analyses.}
}

\references{
  Van Buuren, S. & Oudshoorn, C.G.M. (2000). Multivariate Imputation by Chained Equations: 
  MICE V1.0 User's manual. Report PG/VGZ/00.038, TNO Prevention and Health, Leiden.
}
\author{Stef van Buuren, Karin Oudshoorn, 2000}     
\keyword{misc}  
