% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert.factor.R
\name{mm.column.variable.map}
\alias{mm.column.variable.map}
\title{Define a column-variable map for the model.matrix}
\usage{
mm.column.variable.map(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A character vector indicating to which variable of the original data.frame the model.matrix columns belong

character vector
}
\description{
Given a data.frame for which a model.matrix should be produced, this function returns a vector clearly defining which columns of the model.matrix represent which variable of the original data.frame
}
\details{
This function is used to create an attribute object that the revert.factor function can use to reconstruct the original data.frame from a model.matrix.
}
\author{
Edoardo Costantini, 2023
}
