% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnar_demo_data.R
\docType{data}
\name{mnar_demo_data}
\alias{mnar_demo_data}
\title{MNAR demo data}
\format{
An object of class \code{data.frame} with 500 rows and 3 columns.
}
\source{
\url{https://github.com/moreno-betancur/NARFCS/blob/master/datmis.csv}
}
\usage{
mnar_demo_data
}
\description{
A toy example from Margarita Moreno-Betancur for checking NARFCS.
}
\details{
A small dataset with just three columns.
}
\keyword{datasets}
