% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{name.blocks}
\alias{name.blocks}
\title{Name imputation blocks}
\usage{
name.blocks(blocks, prefix = "b")
}
\arguments{
\item{blocks}{List of \eqn{q} character vectors that identifies the
variable names per block. The name of list elements
identify blocks. \code{mice()} will provide default names
(\code{"b1"}, \code{"b2"}, ...) for blocks containing multiple
variables. Variables within a block are imputed as a
block, e.g. by a multivariate imputation method, or
by an iterated version of the same univariate imputation
method. By default each variable is allocated to a
separate block, which is effectively fully conditional
specification (FCS) by univariate models
(variable-by-variable imputation).
All data variables are assigned to a block.
A variable can belong to only one block, so there are
at most \eqn{p} blocks.
See the \code{parcel} argument for an easier alternative to
the \code{blocks} argument.}

\item{prefix}{A character vector of length 1 with the prefix to
be using for naming any unnamed blocks with two or more variables.}
}
\value{
A named list of character vectors with variables names.
}
\description{
This helper function names any unnamed elements in the \code{blocks}
specification. This is a convenience function.
}
\details{
This function will name any unnamed list elements specified in
the optional argument \code{blocks}. Unnamed blocks
consisting of just one variable will be named after this variable.
Unnamed blocks containing more than one variables will be named
by the \code{prefix} argument, padded by an integer sequence
stating at 1.
}
\examples{
blocks <- list(c("hyp", "chl"), AGE = "age", c("bmi", "hyp"), "edu")
name.blocks(blocks)
}
\seealso{
\code{\link[=mice]{mice()}}
}
