% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncc.R
\name{ncc}
\alias{ncc}
\title{Number of complete cases}
\usage{
ncc(x)
}
\arguments{
\item{x}{An `R` object. Currently supported are methods for the
following classes: `mids`, `data.frame` and `matrix`. Also,
`x` can be a vector.}
}
\value{
Number of elements in `x` with complete data.
}
\description{
Calculates the number of complete cases.
}
\examples{
ncc(nhanes) # 13 complete cases
}
\seealso{
[nic()], [cci()]
}
\author{
Stef van Buuren, 2017
}
