% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncc.R
\name{nic}
\alias{nic}
\title{Number of incomplete cases}
\usage{
nic(x)
}
\arguments{
\item{x}{An `R` object. Currently supported are methods for the
following classes: `mids`, `data.frame` and `matrix`. Also,
`x` can be a vector.}
}
\value{
Number of elements in `x` with incomplete data.
}
\description{
Calculates the number of incomplete cases.
}
\examples{
nic(nhanes) # the remaining 12 rows
nic(nhanes[, c("bmi", "hyp")]) # number of cases with incomplete bmi and hyp
}
\seealso{
[ncc()], [cci()]
}
\author{
Stef van Buuren, 2017
}
