% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mids}
\alias{plot.mids}
\title{Plot the trace lines of the MICE algorithm}
\usage{
\method{plot}{mids}(
  x,
  y = NULL,
  theme = mice.theme(),
  layout = c(2, 3),
  type = "l",
  col = 1:10,
  lty = 1,
  ...
)
}
\arguments{
\item{x}{An object of class `mids`}

\item{y}{A formula that specifies which variables, stream and iterations are plotted.
If omitted, all streams, variables and iterations are plotted.}

\item{theme}{The trellis theme to applied to the graphs. The default is `mice.theme()`.}

\item{layout}{A vector of length 2 given the number of columns and rows in the plot.
The default is `c(2, 3)`.}

\item{type}{Parameter `type` of [panel.xyplot()].}

\item{col}{Parameter `col` of [panel.xyplot()].}

\item{lty}{Parameter `lty` of [panel.xyplot()].}

\item{...}{Extra arguments for [xyplot()].}
}
\value{
An object of class `"trellis"`.
}
\description{
Trace line plots portray the value of an estimate
against the iteration number. The estimate can be anything that you can calculate, but
typically are chosen as parameter of scientific interest. The `plot` method for
a `mids` object plots the mean and standard deviation of the imputed (not observed)
values against the iteration number for each of the $m$ replications. By default,
the function plot the development of the mean and standard deviation for each incomplete
variable. On convergence, the streams should intermingle and be free of any trend.
}
\examples{
imp <- mice(nhanes, print = FALSE)
plot(imp, bmi + chl ~ .it | .ms, layout = c(2, 1))
}
\seealso{
[mice()], [`mids()`][mids-class],
[xyplot()]
}
\author{
Stef van Buuren 2011
}
