\name{popmis}
\alias{popmis}
\docType{data}
\title{Hox pupil popularity data with missing popularity scores}
\description{
Hox pupil popularity data with some missing popularity scores}
\usage{data(popmis)}
\format{
  A data frame with 2000 rows and 7 columns:
  \describe{
    \item{\code{pupil}}{Pupil number within school}
    \item{\code{school}}{School number}
    \item{\code{popular}}{Pupil popularity with 848 missing entries}
    \item{\code{sex}}{Pupil gender}
    \item{\code{texp}}{Teacher experience (years)}
    \item{\code{const}}{Constant intercept term}
    \item{\code{teachpop}}{Teacher popularity}
  }
}

\details{
The original, complete dataset was generated by Joop Hox as an example of 
well-behaved multilevel data set. The distributed data contains missing data 
in pupil popularity.
}

\source{
Hox, J. J. (2002) \emph{Multilevel analysis. Techniques and applications.} 
Mahwah, NJ: Lawrence Erlbaum.
}

\examples{
popmis[1:3,]
}

\keyword{datasets}

