% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-main.R
\name{predPreyRatio}
\alias{predPreyRatio}
\title{Calculate Prey-Predator sizes ratios based on a regression model}
\usage{
predPreyRatio(predType, Linf, thr = NULL, smallest = 0.1)
}
\arguments{
\item{predType}{The type of predation interaction (piscivorous, planktivorous, predacious).}

\item{Linf}{The maximum length of the species.}

\item{thr}{Lengths where a change in predation behaviour is seen.}

\item{smallest}{The smallest size to calculate the relationship. Default to 0.1 cm.}
}
\value{
A table with the prey-predator sizes ratios.
}
\description{
Calculate Prey-Predator sizes ratios based on a regression model
}
\examples{
predPreyRatio(predType = c("planktivorous", "predacious"), Linf = 30, thr = 15)
}
