% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mads}
\alias{print.mads}
\title{Print a `mads` object}
\usage{
\method{print}{mads}(x, ...)
}
\arguments{
\item{x}{Object of class `mads`}

\item{...}{Other parameters passed down to `print.default()`}
}
\value{
`NULL`
}
\description{
Print a `mads` object
}
\seealso{
[`mads()`][mads-class]
}
