% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mids}
\alias{print.mids}
\alias{print.mira}
\alias{print.mice.anova}
\alias{print.mice.anova.summary}
\title{Print a `mids` object}
\usage{
\method{print}{mids}(x, ...)

\method{print}{mira}(x, ...)

\method{print}{mice.anova}(x, ...)

\method{print}{mice.anova.summary}(x, ...)
}
\arguments{
\item{x}{Object of class `mids`, `mira` or `mipo`}

\item{...}{Other parameters passed down to `print.default()`}
}
\value{
`NULL`

`NULL`

`NULL`

`NULL`
}
\description{
Print a `mids` object

Print a `mira` object

Print a `mice.anova` object

Print a `summary.mice.anova` object
}
\seealso{
[`mids()`][mids-class]

[`mira()`][mira-class]

[mipo()]

[mipo()]
}
