% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipo.r
\name{process_mipo}
\alias{process_mipo}
\title{helper function to process a tidied mipo object}
\usage{
process_mipo(z, x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE)
}
\arguments{
\item{z}{data frame with a tidied version of a coefficient matrix}

\item{x}{an "mipo" object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for logistic regression)}
}
\description{
Adds a confidence interval, and possibly exponentiates, a tidied
object.
}
