% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{remove.rhs.variables}
\alias{remove.rhs.variables}
\title{Remove RHS terms involving specified variable names}
\usage{
remove.rhs.variables(ff, vars)
}
\arguments{
\item{ff}{a formula}

\item{vars}{a vector with varianble names to be removed from rhs}
}
\description{
Remove RHS terms involving specified variable names
}
\details{
If all variable are removed, the function return the intercept only model.
}
\examples{
f1 <- y1 + y2 ~ 1 | z + x1 + x2 + x1 * x2
remove.rhs.variables(f1, c("x1", "z"))

# do not touch lhs
f2 <- bmi + chl + hyp ~ 1 | age
remove.rhs.variables(f2, "bmi")
}
