% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert.factor.R
\name{revert.factors}
\alias{revert.factors}
\title{Revert factors from dummy coding}
\usage{
revert.factors(x)
}
\arguments{
\item{x}{A model.matrix with a contrast attribute}
}
\value{
A version of the input \code{x} where dummy codes are replaced with factors.

data.frame
}
\description{
Given a model.matrix with contrast attribute, this function returns a data.frame with factors replacing the dummy coded binary and categorical predictors
}
\details{
This function is used by any mice.impute.gspcr based imputation method to reconstruct categorical predictors as factors so that PCAmix can use them.
If the input \code{x} has no contrast attribute, it returns the input.
}
\author{
Edoardo Costantini, 2023
}
