% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-main.R
\name{runMICE}
\alias{runMICE}
\title{Run a simulation of the MICE model}
\usage{
runMICE(groups, fleets, environment = NULL, T, ndtPerYear = 4, par = NULL,
  Fmult = 1, prices = NULL, eta_crit = 0.57, niter = 7,
  subDtPerYear = 24, verbose = TRUE, pop = NULL, resources = NULL)
}
\arguments{
\item{groups}{A list containing the information to create the functional groups. See details.}

\item{fleets}{A list containing the information to create the fleets. See details.}

\item{environment}{A list containing the environmental information. One value per time step is required.}

\item{ndtPerYear}{Number of time steps per year.}

\item{par}{A list with parameters values. These parameters take precedence over 'groups' and 'fleets'.}

\item{Fmult}{Fishing multiplier. All the fishing mortalities are multiplied by this value.}

\item{niter}{Number of iterations for the calculation of predation mortality.}

\item{verbose}{Logical, should running messages be produced?}

\item{Mstarv}{Maximum starvation mortality.}

\item{Ystar}{Optimal annual food ration per gram of biomass.}

\item{delta}{Fraction of prey population available to predators.}
}
\value{
A list with the abundance (N), length (L) and biomass (B) of
all the species modeled.
}
\description{
Run a simulation of the MICE model
}
