% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{Squeeze the imputed values to be within specified boundaries.}
\usage{
squeeze(x, bounds = NULL, r = rep.int(TRUE, nrow(x)))
}
\arguments{
\item{x}{A \code{data.table} with \emph{m} columns of imputed values followed
by one column called \code{"row_id"} containing the row index in \code{data}.}

\item{bounds}{A numerical vector of length 2 containing the lower and upper bounds.
By default, there are no bounds.}

\item{r}{A logical vector of length \code{length(x)} that is used to select a
subset rows in \code{x} before calculating automatic bounds.}
}
\value{
A \code{data.table} of the same structure as \code{x}
}
\description{
The \code{squeeze()} function is useful for imputed values that are
constrained to a certain range. It replaces imputed values for
imputations in \code{x} lower than \code{bounds[1]} by \code{bounds[1]},
and replaces values higher than \code{bounds[2]} by \code{bounds[2]}.
}
\note{
The function is designed to be called by the \code{post} argument of the
\code{mice()} function. Since the update to \code{data.table} for the \code{imp}
element of the \code{mids} object, the \code{x} argument should now be a
\code{data.table} instead of a vector.
}
\examples{
ini <- mice(boys, maxit = 0)
meth <- ini$meth

# convert to numeric to evade warnings
meth["tv"] <- "norm"
boys$tv <- as.numeric(boys$tv)

post <- ini$post
post["tv"] <- "imp[[j]] <- squeeze(imp[[j]], c(1, 25))"

# without post-processing
imp1 <- mice(boys, meth = meth, print = FALSE, m = 2, maxit = 2, seed = 1)

# with post-processing
imp2 <- mice(boys, meth = meth, post = post, print = FALSE, m = 2,
             maxit = 2, seed = 1)

summary(imp1$imp$tv)
summary(imp2$imp$tv)

oldpar <- par(mfrow = c(1, 2))
plot(imp1$imp$tv[, 1:2], col = mdc(2), xlab = "Imputation 1",
     ylab = "Imputation 2", xlim = c(-20, 30), ylim = c(-20, 30),
     cex = 0.6, main = "No bounds")
plot(imp2$imp$tv[, 1:2], col = mdc(2), xlab = "Imputation 1",
      ylab = "Imputation 2", xlim = c(-20, 30), ylim = c(-20, 30),
      cex = 0.6, main = "Bounds 1 and 25")
par(oldpar)
}
\author{
Stef van Buuren, 2011, 2025
}
