% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripplot.R
\name{stripplot.mids}
\alias{stripplot.mids}
\alias{stripplot}
\title{Stripplot of observed and imputed data}
\usage{
\method{stripplot}{mids}(
  x,
  data,
  na.groups = NULL,
  groups = NULL,
  as.table = TRUE,
  theme = mice.theme(),
  allow.multiple = TRUE,
  outer = TRUE,
  drop.unused.levels = lattice::lattice.getOption("drop.unused.levels"),
  panel = lattice::lattice.getOption("panel.stripplot"),
  default.prepanel = lattice::lattice.getOption("prepanel.default.stripplot"),
  jitter.data = TRUE,
  horizontal = FALSE,
  ...,
  subscripts = TRUE,
  subset = TRUE
)
}
\arguments{
\item{x}{A `mids` object, typically created by `mice()` or
`mice.mids()`.}

\item{data}{Formula that selects the data to be plotted.  This argument
follows the \pkg{lattice} rules for *formulas*, describing the primary
variables (used for the per-panel display) and the optional conditioning
variables (which define the subsets plotted in different panels) to be used
in the plot.

The formula is evaluated on the complete data set in the `long` form.
Legal variable names for the formula include `names(x$data)` plus the
two administrative factors `.imp` and `.id`.

**Extended formula interface:** The primary variable terms (both the LHS
`y` and RHS `x`) may consist of multiple terms separated by a
\sQuote{+} sign, e.g., `y1 + y2 ~ x | a * b`.  This formula would be
taken to mean that the user wants to plot both `y1 ~ x | a * b` and
`y2 ~ x | a * b`, but with the `y1 ~ x` and `y2 ~ x` in
*separate panels*. This behavior differs from standard \pkg{lattice}.
*Only combine terms of the same type*, i.e. only factors or only
numerical variables. Mixing numerical and categorical data occasionally
produces odds labeling of vertical axis.

For convenience, in `stripplot()` and `bwplot` the formula
`y~.imp` may be abbreviated as `y`. This applies only to a single
`y`, and does not (yet) work for `y1+y2~.imp`.}

\item{na.groups}{An expression evaluating to a logical vector indicating
which two groups are distinguished (e.g. using different colors) in the
display. The environment in which this expression is evaluated in the
response indicator `is.na(x$data)`.

The default `na.group = NULL` contrasts the observed and missing data
in the LHS `y` variable of the display, i.e. groups created by
`is.na(y)`. The expression `y` creates the groups according to
`is.na(y)`. The expression `y1 & y2` creates groups by
`is.na(y1) & is.na(y2)`, and `y1 | y2` creates groups as
`is.na(y1) | is.na(y2)`, and so on.}

\item{groups}{This is the usual `groups` arguments in \pkg{lattice}. It
differs from `na.groups` because it evaluates in the completed data
`data.frame(complete(x, "long", inc=TRUE))` (as usual), whereas
`na.groups` evaluates in the response indicator. See
[xyplot()] for more details. When both `na.groups` and
`groups` are specified, `na.groups` takes precedence, and
`groups` is ignored.}

\item{as.table}{See [lattice::xyplot()].}

\item{theme}{A named list containing the graphical parameters. The default
function `mice.theme` produces a short list of default colors, line
width, and so on. The extensive list may be obtained from
`trellis.par.get()`. Global graphical parameters like `col` or
`cex` in high-level calls are still honored, so first experiment with
the global parameters. Many setting consists of a pair. For example,
`mice.theme` defines two symbol colors. The first is for the observed
data, the second for the imputed data. The theme settings only exist during
the call, and do not affect the trellis graphical parameters.}

\item{allow.multiple}{See [lattice::xyplot()].}

\item{outer}{See [lattice::xyplot()].}

\item{drop.unused.levels}{See [lattice::xyplot()].}

\item{panel}{See [xyplot()].}

\item{default.prepanel}{See [lattice::xyplot()].}

\item{jitter.data}{See [lattice::panel.xyplot()].}

\item{horizontal}{See [lattice::xyplot()].}

\item{\dots}{Further arguments, usually not directly processed by the
high-level functions documented here, but instead passed on to other
functions.}

\item{subscripts}{See [lattice::xyplot()].}

\item{subset}{See [lattice::xyplot()].}
}
\value{
The high-level functions documented here, as well as other high-level
Lattice functions, return an object of class `"trellis"`.  The
[`update()`][lattice::update.trellis] method can be used to
subsequently update components of the object, and the
[`print()`][lattice::print.trellis] method (usually called by default)
will plot it on an appropriate plotting device.
}
\description{
Plotting methods for imputed data using \pkg{lattice}.
`stripplot` produces one-dimensional
scatterplots. The function
automatically separates the observed and imputed data. The
functions extend the usual features of \pkg{lattice}.
}
\details{
The argument `na.groups` may be used to specify (combinations of)
missingness in any of the variables. The argument `groups` can be used
to specify groups based on the variable values themselves. Only one of both
may be active at the same time. When both are specified, `na.groups`
takes precedence over `groups`.

Use the `subset` and `na.groups` together to plots parts of the
data. For example, select the first imputed data set by by
`subset=.imp==1`.

Graphical parameters like `col`, `pch` and `cex` can be
specified in the arguments list to alter the plotting symbols. If
`length(col)==2`, the color specification to define the observed and
missing groups. `col[1]` is the color of the 'observed' data,
`col[2]` is the color of the missing or imputed data. A convenient color
choice is `col=mdc(1:2)`, a transparent blue color for the observed
data, and a transparent red color for the imputed data. A good choice is
`col=mdc(1:2), pch=20, cex=1.5`. These choices can be set for the
duration of the session by running `mice.theme()`.
}
\note{
The first two arguments (`x` and `data`) are reversed
compared to the standard Trellis syntax implemented in \pkg{lattice}. This
reversal was necessary in order to benefit from automatic method dispatch.

In \pkg{mice} the argument `x` is always a `mids` object, whereas
in \pkg{lattice} the argument `x` is always a formula.

In \pkg{mice} the argument `data` is always a formula object, whereas in
\pkg{lattice} the argument `data` is usually a data frame.

All other arguments have identical interpretation.
}
\examples{
imp <- mice(boys, maxit = 1)

### stripplot, all numerical variables
\dontrun{
stripplot(imp)
}

### same, but with improved display
\dontrun{
stripplot(imp, col = c("grey", mdc(2)), pch = c(1, 20))
}

### distribution per imputation of height, weight and bmi
### labeled by their own missingness
\dontrun{
stripplot(imp, hgt + wgt + bmi ~ .imp,
  cex = c(2, 4), pch = c(1, 20), jitter = FALSE,
  layout = c(3, 1)
)
}

### same, but labeled with the missingness of wgt (just four cases)
\dontrun{
stripplot(imp, hgt + wgt + bmi ~ .imp,
  na = wgt, cex = c(2, 4), pch = c(1, 20), jitter = FALSE,
  layout = c(3, 1)
)
}

### distribution of age and height, labeled by missingness in height
### most height values are missing for those around
### the age of two years
### some additional missings occur in region WEST
\dontrun{
stripplot(imp, age + hgt ~ .imp | reg, hgt,
  col = c(grDevices::hcl(0, 0, 40, 0.2), mdc(2)), pch = c(1, 20)
)
}

### heavily jitted relation between two categorical variables
### labeled by missingness of gen
### aggregated over all imputed data sets
\dontrun{
stripplot(imp, gen ~ phb, factor = 2, cex = c(8, 1), hor = TRUE)
}

### circle fun
stripplot(imp, gen ~ .imp,
  na = wgt, factor = 2, cex = c(8.6),
  hor = FALSE, outer = TRUE, scales = "free", pch = c(1, 19)
)
}
\references{
Sarkar, Deepayan (2008) *Lattice: Multivariate Data
Visualization with R*, Springer.

van Buuren S and Groothuis-Oudshoorn K (2011). `mice`: Multivariate
Imputation by Chained Equations in `R`. *Journal of Statistical
Software*, **45**(3), 1-67. \doi{10.18637/jss.v045.i03}
}
\seealso{
[mice()], [xyplot()], [densityplot()],
[bwplot()], [lattice()] for an overview of the
package, as well as [`stripplot()`][lattice::xyplot],
[lattice::panel.stripplot()],
[lattice::print.trellis()],
[`trellis.par.set()`][lattice::trellis.par.get]
}
\author{
Stef van Buuren
}
\keyword{hplot}
