% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.transparent.R
\name{supports.transparent}
\alias{supports.transparent}
\alias{transparent}
\title{Supports semi-transparent foreground colors?}
\usage{
supports.transparent()
}
\value{
`TRUE` or `FALSE`
}
\description{
This function is used by `mdc()` to find out whether the current device
supports semi-transparent foreground colors.
}
\details{
The function calls the function `dev.capabilities()` from the package
`grDevices`. The function return `FALSE` if the status of the
current device is unknown.
}
\examples{

supports.transparent()
}
\seealso{
[mdc()] [dev.capabilities()]
}
\keyword{hplot}
