% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.mipo}
\alias{tidy.mipo}
\title{Tidy method to extract results from a \code{mipo} object}
\usage{
\method{tidy}{mipo}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{mipo}}

\item{conf.int}{Logical. Should confidence intervals be returned?}

\item{conf.level}{Confidence level for intervals. Defaults to .95}

\item{...}{extra arguments (not used)}
}
\value{
A dataframe withh these columns:
\itemize{
\item term
\item estimate
\item ubar
\item b
\item t
\item dfcom
\item df
\item riv
\item lambda
\item fmi
\item p.value
\item conf.low (if called with conf.int = TRUE)
\item conf.high (if called with conf.int = TRUE)
}
}
\description{
Tidy method to extract results from a \code{mipo} object
}
\keyword{internal}
