% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.mira}
\alias{tidy.mira}
\title{Tidy method to extract results from a `mira` object}
\usage{
\method{tidy}{mira}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{mira}}

\item{conf.int}{Logical. Should confidence intervals be returned?}

\item{conf.level}{Confidence level for intervals. Defaults to .95}

\item{...}{extra arguments (not used)}
}
\value{
A dataframe withh these columns:
\itemize{
     \item term
     \item estimate
     \item ubar
     \item b
     \item t
     \item dfcom
     \item df
     \item riv
     \item lambda
     \item fmi
     \item p.value
     \item conf.low (if called with conf.int = TRUE)
     \item conf.high (if called with conf.int = TRUE)
}
}
\description{
Tidy method to extract results from a `mira` object
}
\keyword{internal}
