% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.mads.R
\name{xyplot.mads}
\alias{xyplot.mads}
\title{Scatterplot of amputed and non-amputed data against weighted sum scores}
\usage{
\method{xyplot}{mads}(x, yvar = NULL, which.pat = NULL,
  standardized = TRUE, layout = NULL, theme = mice.theme(transparent =
  FALSE), multiple = TRUE, outer = TRUE)
}
\arguments{
\item{x}{A \code{mads} object, typically created by \code{\link{ampute}}.}

\item{yvar}{A string or vector of variable names that needs to be plotted. As 
a default, all variables will be plotted.}

\item{which.pat}{A scalar or vector indicating which patterns need to be plotted. 
As a default, all patterns are plotted.}

\item{standardized}{Logical. Whether the scatterplots need to be created
from standardized data or not. Default is TRUE.}

\item{layout}{A vector of two values indicating how the scatterplots of one 
pattern should be divided over the plot. For example, \code{c(2, 3)} indicates 
that the scatterplots of six variables need to be placed on 3 rows and 2 columns. 
There are several defaults for different #variables. Note that for more than 
9 variables, multiple plots will be created automatically.}

\item{theme}{A named list containing the graphical parameters. The default
function \code{mice.theme} produces a short list of default colors, line
width, and so on. The extensive list may be obtained from
\code{trellis.par.get}.}

\item{multiple}{See \code{\link[lattice:bwplot]{bwplot}}.}

\item{outer}{See \code{\link[lattice:bwplot]{bwplot}}.}
}
\value{
A list containing the scatterplots. Note that a new pattern 
will always be shown in a new plot.
}
\description{
Plotting method for amputed data using \pkg{lattice}. \code{xyplot}
produces scatterplots. The function plots the variables against the weighted 
sum scores. This is useful in analyzing the missing data mechanisms. The function
automatically separates the amputed and non-amputed data.
}
\note{
The \code{mads} object contains all the information you need to 
make any desired plots. Check \code{\link{mads-class}} or the vignette titled "Multivariate 
Amputation using Ampute" to understand the contents of class object \code{mads}.
}
\author{
Rianne Schouten, 2016
}
\seealso{
\code{\link{ampute}}, \code{\link{bwplot}}, \code{\link{Lattice}} for 
an overview of the package, \code{\link{mads-class}}
}

