% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampute.default.R
\name{ampute.default.patterns}
\alias{ampute.default.patterns}
\title{Default \code{patterns} in \code{ampute}}
\usage{
ampute.default.patterns(n)
}
\arguments{
\item{n}{A scalar specifying the #variables in the data.}
}
\value{
A square matrix of size #variables where \code{0} indicates a variable
}
\description{
This function creates a default pattern matrix for the multivariate 
amputation function \code{ampute()}.
}
\author{
Rianne Schouten, 2016
}
\seealso{
\code{\link{ampute}}, \code{\link{md.pattern}}
}

