% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mira}
\alias{anova.mira}
\title{Compare several nested models}
\usage{
\method{anova}{mira}(object, ..., method = "D1", use = "wald")
}
\arguments{
\item{object}{Two or more objects of class \code{mira}}

\item{...}{Other parameters passed down to \code{D1()}, \code{D2()},
\code{D3()} and \code{mitml::testModels}.}

\item{method}{Either \code{"D1"}, \code{"D2"} or \code{"D3"}}

\item{use}{An character indicating the test statistic}
}
\value{
Object of class \code{mice.anova}
}
\description{
Compare several nested models
}
