% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Computes convergence diagnostics for a \code{mids} object}
\usage{
convergence(data, diagnostic = "all", ...)
}
\arguments{
\item{data}{An object of class \code{mids} as created by the function
\code{mice()}.}

\item{diagnostic}{A keyword. One of the following keywords: \code{"ac"},
\code{"all"}, \code{"gr"} and \code{"psrf"}. See the Details section
for the interpretation.
The default is \code{diagnostic = "all"} which returns both the
autocorrelation and potential scale reduction factor per iteration.}

\item{\dots}{Additional arguments. Not used.}
}
\value{
A \code{data.frame} with the autocorrelation and/or potential
scale reduction factor per iteration of the MICE algorithm.
}
\description{
Takes an object of class \code{mids}, computes the autocorrelation
and/or potential scale reduction factor, and returns a \code{data.frame}
with the specified diagnostic(s) per iteration.
}
\details{
The argument \code{diagnostic} can be length-1 character, which is
matched to one of the following keywords:
\describe{
\item{\code{"all"}}{computes both the autocorrelation as well as the
potential scale reduction factor per iteration;}
\item{\code{"ac"}}{computes only the autocorrelation per iteration;}
\item{\code{"psrf"}}{computes only the potential scale reduction factor
per iteration;}
\item{\code{"gr"}}{same as \code{psrf}, the potential scale reduction
factor is colloquially called the Gelman-Rubin diagnostic.}
}
}
\examples{

# obtain imputed data set
imp <- mice(nhanes2, print = FALSE)
# compute convergence diagnostics
convergence(imp)
}
\seealso{
\code{\link{mice}}, \code{\link[=mids-class]{mids}}
}
\keyword{none}
