% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1.R
\name{D1}
\alias{D1}
\title{Compare two nested models using D1-statistic}
\usage{
D1(fit1, fit0 = NULL, dfcom = NULL, df.com = NULL)
}
\arguments{
\item{fit1}{An object of class `mira`, produced by `with()`.}

\item{fit0}{An object of class `mira`, produced by `with()`. The
model in `fit0` is a nested within `fit1`. The default null
model `fit0 = NULL` compares `fit1` to the intercept-only model.}

\item{dfcom}{A single number denoting the
complete-data degrees of freedom of model `fit1`. If not specified,
it is set equal to `df.residual` of model `fit1`. If that cannot
be done, the procedure assumes (perhaps incorrectly) a large sample.}

\item{df.com}{Deprecated}
}
\description{
The D1-statistics is the multivariate Wald test.
}
\note{
Warning: `D1()` assumes that the order of the variables is the
same in different models. See
<https://github.com/amices/mice/issues/420> for details.
}
\examples{
# Compare two linear models:
imp <- mice(nhanes2, seed = 51009, print = FALSE)
mi1 <- with(data = imp, expr = lm(bmi ~ age + hyp + chl))
mi0 <- with(data = imp, expr = lm(bmi ~ age + hyp))
D1(mi1, mi0)
\dontrun{
# Compare two logistic regression models
imp <- mice(boys, maxit = 2, print = FALSE)
fit1 <- with(imp, glm(gen > levels(gen)[1] ~ hgt + hc + reg, family = binomial))
fit0 <- with(imp, glm(gen > levels(gen)[1] ~ hgt + hc, family = binomial))
D1(fit1, fit0)
}
}
\references{
Li, K. H., T. E. Raghunathan, and D. B. Rubin. 1991.
Large-Sample Significance Levels from Multiply Imputed Data Using
Moment-Based Statistics and an F Reference Distribution.
*Journal of the American Statistical Association*, 86(416): 1065–73.

<https://stefvanbuuren.name/fimd/sec-multiparameter.html#sec:wald>
}
\seealso{
[mitml::testModels()]
}
