% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.mids.R
\name{filter.mids}
\alias{filter.mids}
\title{Filter data and imputations contained in a \code{mids} object}
\usage{
filter.mids(obj, include = NULL)
}
\arguments{
\item{obj}{A \code{mids} object.}

\item{include}{A logical \code{vector} of same length as \code{nrow(obj$data)}.}
}
\value{
An S3 object of class \code{mids}
}
\description{
This function takes a \code{mids} object and returns a new
\code{mids} object that pertains to the subset of the data
identified by \code{include}.
}
\note{
The function construct the elements of the new \code{mids} object as follows:
\tabular{ll}{
\code{data}     \tab Select rows in \code{obj$data} for which \code{include == TRUE}\cr
\code{imp}      \tab Select rows each imputation \code{data.frame} in \code{obj$imp} for which \code{include == TRUE}\cr
\code{m}        \tab Equals \code{obj$m}\cr
\code{where}    \tab Select rows in \code{obj$where} for which \code{include == TRUE}\cr
\code{blocks}   \tab Equals \code{obj$blocks}\cr
\code{call}     \tab Equals \code{obj$call}\cr
\code{nmis}     \tab Recalculate \code{nmis} based on the selected \code{data} rows\cr
\code{method}   \tab Equals \code{obj$method}\cr
\code{predictorMatrix} \tab Equals \code{obj$predictorMatrix}\cr
\code{visitSequence}   \tab Equals \code{obj$visitSequence}\cr
\code{formulas}  \tab Equals \code{obj$formulas}\cr
\code{post}      \tab Equals \code{obj$post}\cr
\code{blots}     \tab Equals \code{obj$blots}\cr
\code{ignore}    \tab Select positions in \code{obj$ignore} for which \code{include == TRUE}\cr
\code{seed}            \tab Equals \code{obj$seed}\cr
\code{iteration}       \tab Equals \code{obj$iteration}\cr
\code{lastSeedValue}   \tab Equals \code{obj$lastSeedValue}\cr
\code{chainMean}       \tab Set to \code{NULL}\cr
\code{chainVar}        \tab Set to \code{NULL}\cr
\code{loggedEvents}    \tab Equals \code{obj$loggedEvents}\cr
\code{version}    \tab Replaced with current version\cr
\code{date}       \tab Replaced with current date
}
}
\examples{
imp <- mice(nhanes, m = 2, maxit = 1, print = FALSE)
imp_f <- filter.mids(imp, include = c(rep(TRUE, 13), rep(FALSE, 12)))

nrow(complete(imp))
nrow(complete(imp_f))
}
\author{
Patrick Rockenschaub
}
\keyword{manip}
