% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.coef.R
\name{fix.coef}
\alias{fix.coef}
\title{Fix coefficients and update model}
\usage{
fix.coef(model, beta = NULL)
}
\arguments{
\item{model}{An R model, e.g., produced by \code{lm} or \code{glm}}

\item{beta}{A numeric vector with \code{length(coef)} model coefficients. 
If the vector is not named, the coefficients should be
given in the same order as in \code{coef(model)}. If the vector is named,
the procedure attempts to match on names.}
}
\value{
An updated R model object
}
\description{
Refits a model with a specified set of coefficients.
}
\details{
The function calculates the linear predictor using the new coefficients, 
and reformulates the model using the \code{offset} 
argument. The linear predictor is called 
\code{offset}, and its coefficient will be \code{1} by definition. 
The new model only fits the intercept, which should be \code{0} 
if we set \code{beta = coef(model)}.
}
\examples{
fit0 <- lm(Volume ~ Girth + Height, data = trees)
formula(fit0)
coef(fit0)
deviance(fit0)

# refit same model
fit1 <- fix.coef(fit0)
formula(fit1)
coef(fit1)
deviance(fit1)

# change the beta's
fit2 <- fix.coef(fit0, beta = c(-50, 5, 1))
coef(fit2)
deviance(fit2)

# compare predictions
plot(predict(fit0), predict(fit1)); abline(0,1)
plot(predict(fit0), predict(fit2)); abline(0,1)

# compare proportion explained variance
cor(predict(fit0), predict(fit0) + residuals(fit0))^2
cor(predict(fit1), predict(fit1) + residuals(fit1))^2
cor(predict(fit2), predict(fit2) + residuals(fit2))^2

# extract offset from constrained model
summary(fit2$model$offset)

# it also works with factors and missing data
fit0 <- lm(bmi ~ age + hyp + chl, data = nhanes2)
fit1 <- fix.coef(fit0)
fit2 <- fix.coef(fit0, beta = c(15, -8, -8, 2, 0.2))
}
\author{
Stef van Buuren, 2018
}
