% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfit.R
\name{getqbar}
\alias{getqbar}
\title{Extract estimate from \code{mipo} object}
\usage{
getqbar(x)
}
\arguments{
\item{x}{An object of class \code{mipo}}
}
\description{
\code{getqbar} returns a named vector of pooled estimates.
}
