% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{ifdo}
\alias{ifdo}
\title{Conditional imputation helper}
\usage{
ifdo(cond, action)
}
\arguments{
\item{cond}{a condition}

\item{action}{the action to do}
}
\value{
Currently returns an error message.
}
\description{
Sorry, the \code{ifdo()} function is not yet implemented.
}
\author{
Stef van Buuren, 2012
}
\keyword{internal}
