% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{ifdo}
\alias{ifdo}
\title{Conditional imputation helper}
\usage{
ifdo(cond, action)
}
\arguments{
\item{cond}{}

\item{action}{}
}
\value{
Currently returns an error message.
}
\description{
Sorry, the \code{ifdo()} function is not yet implemented.
}
\author{
Stef van Buuren, 2012
}
\keyword{internal}

