% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.r
\name{is.mads}
\alias{is.mads}
\title{Check for \code{mads} object}
\usage{
is.mads(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether \code{x} is an object of class \code{mads}
}
\description{
Check for \code{mads} object
}

