% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.mids}
\alias{is.mids}
\title{Check for \code{mids} object}
\usage{
is.mids(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether \code{x} is an object of class \code{mids}
}
\description{
Check for \code{mids} object
}

