% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mids}
\alias{is.mids}
\title{Check for \code{mids} object}
\usage{
is.mids(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether \code{x} is an object of class \code{mids}
}
\description{
Check for \code{mids} object
}
