% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.r
\name{is.mipo}
\alias{is.mipo}
\title{Check for \code{mipo} object}
\usage{
is.mipo(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether \code{x} is an object of class \code{mipo}
}
\description{
Check for \code{mipo} object
}

