% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.mira}
\alias{is.mira}
\title{Check for \code{mira} object}
\usage{
is.mira(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating whether \code{x} is an object of class \code{mira}
}
\description{
Check for \code{mira} object
}

