% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.predictors.R
\name{lars.filter}
\alias{lars.filter}
\title{Filter out predictors by least angular regression}
\usage{
lars.filter(x, y, ry, eps = 1e-04, allow.na = TRUE, ...)
}
\arguments{
\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{eps}{numeric. Criteriom for removing predictors with zero variance.
If \code{eps} is zero, the filter is bypassed. The default is 1e-04.
Not passed down to \code{lars::lars}.}

\item{allow.na}{logical. If \code{TRUE}, allow imputation of fully
missing \code{y}. This typically only occurs for passive imputation.
The default is \code{TRUE}.}

\item{...}{further arguments passed to \code{lars::lars}, like standard
arguments \code{type} and \code{intercept}, or custom tuning parameters
like \code{lars.relax} and \code{minimal.cp}.}
}
\value{
A logical vector of length \code{ncol(x)} indicating which predictors
to keep.
}
\description{
Filter out predictors by least angular regression
}
\details{
In the following conditions, the function skips the filter:
1. if there are no or only one predictors
2. if user imputes a fully missing y accept all predictors
3. if eps == 0 (for backward compatibility)
}
