\name{long2mids}
\alias{long2mids}
\title{Conversion of a imputed data set (long form) to a \code{mids} object}
\usage{
  long2mids(x)
}
\arguments{
  \item{x}{}
}
\value{
  An object of class \code{mids}
}
\description{
  Conversion of a imputed data set (long form) to a
  \code{mids} object
}
\author{
  Stef van Buuren, 2012
}
\keyword{internal}

