% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{make.post}
\alias{make.post}
\title{Creates a \code{post} argument}
\usage{
make.post(data)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.  Missing
values are coded as \code{NA}.}
}
\value{
Character vector of \code{ncol(data)} element
}
\description{
This helper function creates a valid \code{post} vector. The
\code{post} vector is an argument to the \code{mice} function that
specifies post-processing for a variable just after imputation.
}
\examples{
make.post(nhanes2)
}
\seealso{
\code{\link{mice}}
}
