% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitSequence.R
\name{make.visitSequence}
\alias{make.visitSequence}
\title{Creates a \code{visitSequence} argument}
\usage{
make.visitSequence(data = NULL, blocks = NULL)
}
\arguments{
\item{data}{Data frame with \eqn{n} rows and \eqn{p} columns with
incomplete data.  Missing values are coded as \code{NA}.}

\item{blocks}{List of \eqn{q} character vectors that identifies the
variable names per block. The name of list elements
identify blocks. \code{mice()} will provide default names
(\code{"B1"}, \code{"B2"}, ...) for blocks containing multiple
variables. Variables within a block are imputed as a
block, e.g. by a multivariate imputation method, or
by an iterated version of the same univariate imputation
method. By default each variable is allocated to a
separate block, which is effectively fully conditional
specification (FCS) by univariate models
(variable-by-variable imputation).
All data variables are assigned to a block.
A variable can belong to only one block, so there are
at most \eqn{p} blocks.
See the \code{nest} argument for an easier alternative to
the \code{blocks} argument.}
}
\value{
Vector containing block names
}
\description{
This helper function creates a valid \code{visitSequence}. The
\code{visitSequence} is an argument to the \code{mice} function that
specifies the sequence in which blocks are imputed.
}
\examples{
make.visitSequence(nhanes)
}
\seealso{
\code{\link[=mice]{mice()}}
}
