\name{mice.auxiliary}

\alias{long2mids}
\alias{ifdo}
\alias{appendbreak}
\alias{extractBS}

\title{Auxiliary functions used in FIMD}

\description{
Auxiliary function that are not (yet) part of \code{mice}, but that are needed to execute the code for the 
book Flexible Imputation of Missing Data (FIMD).
}

\usage{
long2mids(x)
ifdo(cond, action)
appendbreak(data, brk, warp.model = warp.model, id=NULL, typ="pred")
extractBS(fit)
}

\arguments{
    \item{x}{}
    \item{cond}{}
    \item{action}{}
    \item{data}{}
    \item{brk}{}
    \item{warp.model}{}
    \item{id}{}
    \item{typ}{}
    \item{fit}{}
}

\author{Stef van Buuren, 2012} 

\keyword{internal}  
