% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.bin.R
\name{mice.impute.2l.bin}
\alias{mice.impute.2l.bin}
\title{Imputation by a two-level logistic model using `glmer`}
\usage{
mice.impute.2l.bin(y, ry, x, type, wy = NULL, intercept = TRUE, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{type}{Vector of length `ncol(x)` identifying random and class
variables.  Random variables are identified by a '2'. The class variable
(only one is allowed) is coded as '-2'. Fixed effects are indicated by
a '1'.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{intercept}{Logical determining whether the intercept is automatically
added.}

\item{\dots}{Arguments passed down to `glmer`}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Imputes univariate systematically and sporadically missing data
using a two-level logistic model using `lme4::glmer()`
}
\details{
Data are missing systematically if they have not been measured, e.g., in the
case where we combine data from different sources. Data are missing sporadically
if they have been partially observed.
}
\examples{
library(tidyr)
library(dplyr)
data("toenail2")
data <- tidyr::complete(toenail2, patientID, visit) \%>\%
  tidyr::fill(treatment) \%>\%
  dplyr::select(-time) \%>\%
  dplyr::mutate(patientID = as.integer(patientID))
\dontrun{
pred <- mice(data, print = FALSE, maxit = 0, seed = 1)$pred
pred["outcome", "patientID"] <- -2
imp <- mice(data, method = "2l.bin", pred = pred, maxit = 1, m = 1, seed = 1)
}
}
\references{
Jolani S., Debray T.P.A., Koffijberg H., van Buuren S., Moons K.G.M. (2015).
Imputation of systematically missing predictors in an individual
participant data meta-analysis: a generalized approach using MICE.
*Statistics in Medicine*, 34:1841-1863.
}
\seealso{
Other univariate-2l: 
\code{\link{mice.impute.2l.lmer}()},
\code{\link{mice.impute.2l.norm}()},
\code{\link{mice.impute.2l.pan}()}
}
\author{
Shahab Jolani, 2015; adapted to mice, SvB, 2018
}
\concept{univariate-2l}
\keyword{datagen}
