% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pcovr.R
\name{mice.impute.pcovr}
\alias{mice.impute.pcovr}
\alias{pcovr}
\title{Imputation by principal covariates regression}
\usage{
mice.impute.pcovr(y, ry, x, wy = NULL, npcs = 1L, DoF = "naive", ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{npcs}{The number of principal components to extract for PC regression.}

\item{DoF}{The method to compute the degrees of freedom, either \code{"naive"}, or \code{"kramer"}.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data by using principal covariates regression (PCovR).
}
\details{
Imputation of \code{y} by principal covariates regression (De Jong and Kiers, 1992).
The method consists of the following steps:
\enumerate{
\item For a given \code{y} variable under imputation, draw a bootstrap version y*
with replacement from the observed cases \code{y[ry]}, and stores in x* the
corresponding values from \code{x[ry, ]}.
\item Compute an optimal \code{alpha} value for the PCovR model on the basis 
of maximum likelihood principles described in Vervloet et. al. (2015) for a given number of components.
\item Estimate W and Py from the PCovR model fitted to y* and x*
\item Predict y values based on \code{x[wy,]} and add noise.
}

Residual degrees of freedom are estimated by using the Krylov representation
described by Kramer and Sugiyama (2011) as implemented in the \code{plsdof} 
package, or with the naive approach (n - npcs - 1).
}
\references{
De Jong, S., & Kiers, H. A. (1992). Principal covariates regression: part I. 
Theory. Chemometrics and Intelligent Laboratory Systems, 14(1-3), 155-164.

Krämer, N., & Sugiyama, M. (2011). The degrees of freedom of partial least
squares regression. Journal of the American Statistical Association, 106(494), 697-705.

Vervloet, M., Kiers, H. A., Van den Noortgate, W., & Ceulemans, E. (2015). 
PCovR: An R package for principal covariates regression. Journal of Statistical Software, 65, 1-14.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pcr}()},
\code{\link{mice.impute.plsr}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()},
\code{\link{mice.impute.spcr}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
