% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pls.R
\name{mice.impute.pls}
\alias{mice.impute.pls}
\alias{pls}
\title{Imputation by partial least square regression}
\usage{
mice.impute.pls(y, ry, x, wy = NULL, nlvs = 1L, DoF = "kramer", ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{nlvs}{The number of latent variables to consider in the PLS regression.}

\item{DoF}{The method to compute the degrees of freedom, either \code{"naive"}, or \code{"kramer"}.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data by using partial least square regression.
}
\details{
Imputation of \code{y} by partial least square regression (Wold, 1975).
The method consists of the following steps:
\enumerate{
\item For a given \code{y} variable under imputation, draw a bootstrap version y*
with replacement from the observed cases \code{y[ry]}, and stores in x* the
corresponding values from \code{x[ry, ]}.
\item PLSR is estimated on this sample for a given number of latent variables (nlvs).
\item Based on this PLSR model, predictions on the new data x[wy, ] are returned
\item Imputations are then obtained by adding noise to these predictions.
}

Noise is added by sampling from a standard normal distribution rescaled based 
on the residuals standard error of the PLSR model.
The residual degrees of freedom are computed based on the Krylov representation 
described by Kramer and Sugiyama (2011) as implemented in the R package \code{plsdof} 
(Krämer and Sugiyama, 2011), or with the naive approach (n - nlvs - 1).
}
\references{
Krämer, N., & Sugiyama, M. (2011). The degrees of freedom of partial least 
squares regression. Journal of the American Statistical Association, 106(494), 697-705.

Wold, H. (1975). Path models with latent variables: The NIPALS approach. 
In Quantitative sociology (pp. 307-357). Academic Press.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.impute.cart}()},
\code{\link{mice.impute.lasso.logreg}()},
\code{\link{mice.impute.lasso.norm}()},
\code{\link{mice.impute.lasso.select.logreg}()},
\code{\link{mice.impute.lasso.select.norm}()},
\code{\link{mice.impute.lda}()},
\code{\link{mice.impute.logreg.boot}()},
\code{\link{mice.impute.logreg}()},
\code{\link{mice.impute.mean}()},
\code{\link{mice.impute.midastouch}()},
\code{\link{mice.impute.mnar.logreg}()},
\code{\link{mice.impute.mpmm}()},
\code{\link{mice.impute.norm.boot}()},
\code{\link{mice.impute.norm.nob}()},
\code{\link{mice.impute.norm.predict}()},
\code{\link{mice.impute.norm}()},
\code{\link{mice.impute.pcovr}()},
\code{\link{mice.impute.pcr}()},
\code{\link{mice.impute.pmm}()},
\code{\link{mice.impute.polr}()},
\code{\link{mice.impute.polyreg}()},
\code{\link{mice.impute.quadratic}()},
\code{\link{mice.impute.rf}()},
\code{\link{mice.impute.ri}()},
\code{\link{mice.impute.spcr}()}
}
\author{
Edoardo Costantini, 2022
}
\concept{univariate imputation functions}
\keyword{datagen}
