% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mice.impute.polr}
\alias{mice.impute.polr}
\title{Imputation by polytomous regression - ordered}
\usage{
mice.impute.polr(y, ry, x, nnet.maxit = 100, nnet.trace = FALSE,
  nnet.maxNWts = 1500, ...)
}
\arguments{
\item{y}{Incomplete data vector of length \code{n}}

\item{ry}{Vector of missing data pattern (\code{FALSE}=missing,
\code{TRUE}=observed)}

\item{x}{Matrix (\code{n} x \code{p}) of complete covariates.}

\item{nnet.maxit}{Tuning parameter for \code{nnet()}.}

\item{nnet.trace}{Tuning parameter for \code{nnet()}.}

\item{nnet.maxNWts}{Tuning parameter for \code{nnet()}.}

\item{...}{Other named arguments.}
}
\value{
A vector of length \code{nmis} with imputations.
}
\description{
Imputes missing data in a categorical variable using polytomous regression
}
\details{
By default, ordered factors with more than two levels are imputed by
\code{mice.impute.polr}.

The function \code{mice.impute.polr()} imputes for ordered categorical response
variables by the proportional odds logistic regression (polr) model. The
function repeatedly applies logistic regression on the successive splits. The
model is also known as the cumulative link model.

The algorithm of \code{mice.impute.polr} uses the function \code{polr()} from
the \code{MASS} package.

In order to avoid bias due to perfect prediction, the algorithm augment the
data according to the method of White, Daniel and Royston (2010).

The call to \code{polr} might fail, usually because the data are very sparse.
In that case, \code{multinom} is tried as a fallback, and a record is written
to the \code{loggedEvents} component of the \code{mids} object.
}
\author{
Stef van Buuren, Karin Groohuis-Oudshoorn, 2000-2010
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). \code{mice}: Multivariate
Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical
Software}, \bold{45}(3), 1-67. \url{http://www.jstatsoft.org/v45/i03/}

Brand, J.P.L. (1999) \emph{Development, implementation and evaluation of
multiple imputation strategies for the statistical analysis of incomplete
data sets.} Dissertation. Rotterdam: Erasmus University.

White, I.R., Daniel, R. Royston, P. (2010). Avoiding bias due to perfect
prediction in multiple imputation of incomplete categorical variables.
\emph{Computational Statistics and Data Analysis}, 54, 2267-2275.

Venables, W.N. & Ripley, B.D. (2002). \emph{Modern applied statistics with
S-Plus (4th ed)}. Springer, Berlin.
}
\seealso{
\code{\link{mice}}, \code{\link[nnet]{multinom}},
\code{\link[MASS]{polr}}
}
\keyword{datagen}

