% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.sample.R
\name{mice.impute.sample}
\alias{mice.impute.sample}
\title{Imputation by simple random sampling}
\usage{
mice.impute.sample(y, ry, x = NULL, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length `length(y)` indicating the
the subset `y[ry]` of elements in `y` to which the imputation
model is fitted. The `ry` generally distinguishes the observed
(`TRUE`) and missing values (`FALSE`) in `y`.}

\item{x}{Numeric design matrix with `length(y)` rows with predictors for
`y`. Matrix `x` may have no missing values.}

\item{wy}{Logical vector of length `length(y)`. A `TRUE` value
indicates locations in `y` for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as `y`, and of length
`sum(wy)`
}
\description{
Imputes a random sample from the observed `y` data
}
\details{
This function takes a simple random sample from the observed values in
`y`, and returns these as imputations.
}
\references{
van Buuren S and Groothuis-Oudshoorn K (2011). `mice`:
Multivariate Imputation by Chained Equations in `R`. *Journal of
Statistical Software*, **45**(3), 1-67.
\doi{10.18637/jss.v045.i03}
}
\author{
Stef van Buuren, Karin Groothuis-Oudshoorn, 2000, 2017
}
\keyword{datagen}
