% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipo.r
\name{mipo}
\alias{mipo}
\title{\code{mipo}: Multiple imputation pooled object}
\usage{
mipo(object, ...)
}
\arguments{
\item{object}{An object of class \code{mira}.}

\item{\dots}{Passed down to \code{pool}.}
}
\description{
The \code{mipo} object is generated by the function \code{\link{pool}}.
The \code{mipo} class is of objects has methods for
the following generic functions: \code{print}, \code{summary}.
}
\details{
An object class \code{mipo} is a \code{list} containing the 
   following elements:
   
\tabular{ll}{
\code{call}    \tab Call that created the \code{mipo} object\cr
\code{m}       \tab Scalar, number of multiple imputations\cr
\code{term}    \tab Vector, name of the parameter\cr
\code{qbar}    \tab Vector, pooled complete data estimate\cr
\code{ubar}    \tab Vector, within-imputation variance of \code{qbar}\cr
\code{b}       \tab Vector, between-imputation variance of \code{qbar}\cr
\code{t}       \tab Vector, total variance, of \code{qbar}\cr
\code{dfcom}   \tab Vector, degrees of freedom in complete data\cr
\code{df}      \tab Vector, degrees of freedom of $t$-statistic\cr
\code{riv}     \tab Vcetor, relative increase in variance\cr
\code{lambda}  \tab Vector, proportion attributable to the missingness\cr
\code{fmi}     \tab Vector, fraction of missing information\cr
}
}
\references{
van Buuren S and Groothuis-Oudshoorn K (2011). \code{mice}:
Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of
Statistical Software}, \bold{45}(3), 1-67.
\url{http://www.jstatsoft.org/v45/i03/}
}
\seealso{
\code{\link{pool}}, \code{\link[=mids-class]{mids}}, \code{\link[=mira-class]{mira}}
}
\keyword{classes}
