% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncc.R
\name{ncc}
\alias{ncc}
\title{Number of complete cases}
\usage{
ncc(x)
}
\arguments{
\item{x}{An \code{R} object. Currently supported are methods for the
following classes: \code{mids}, \code{data.frame} and \code{matrix}. Also, 
\code{x} can be a vector.}
}
\value{
Number of elements in \code{x} with complete data.
}
\description{
Calculates the number of complete cases.
}
\examples{

 ncc(nhanes) # 13 complete cases 

}
\author{
Stef van Buuren, 2017
}
\seealso{
\code{\link{nic}}, \code{\link{cci}}
}

