\name{nhanes}
\alias{nhanes}
\docType{data}
\title{nhanes data set}
\description{
 
}
\usage{data(nhanes)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{age}{a numeric vector}
    \item{bmi}{a numeric vector}
    \item{hyp}{a numeric vector}
    \item{chl}{a numeric vector}
  }
}

}
\source{
 
}


\keyword{datasets}
