\name{nhanes2}
\alias{nhanes2}
\docType{data}
\title{ nhanes2 data set}
\description{

}
\usage{data(nhanes2)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{age}{a factor with levels \code{1} \code{2} \code{3}}
    \item{bmi}{a numeric vector}
    \item{hyp}{a factor with levels \code{1} \code{2}}
    \item{chl}{a numeric vector}
  }
}

\source{

}

\keyword{datasets}
