% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncc.R
\name{nic}
\alias{nic}
\title{Number of incomplete cases}
\usage{
nic(x)
}
\arguments{
\item{x}{An \code{R} object. Currently supported are methods for the
following classes: \code{mids}, \code{data.frame} and \code{matrix}. Also,
\code{x} can be a vector.}
}
\value{
Number of elements in \code{x} with incomplete data.
}
\description{
Calculates the number of incomplete cases.
}
\examples{
nic(nhanes) # the remaining 12 rows
nic(nhanes[, c("bmi", "hyp")]) # number of cases with incomplete bmi and hyp
}
\seealso{
\code{\link{ncc}}, \code{\link{cci}}
}
\author{
Stef van Buuren, 2017
}
