% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmis.R
\docType{data}
\name{popmis}
\alias{popmis}
\title{Hox pupil popularity data with missing popularity scores}
\format{
A data frame with 2000 rows and 7 columns: 
\describe{
\item{pupil}{Pupil number within school} 
\item{school}{School number} 
\item{popular}{Pupil popularity with 848 missing entries}
\item{sex}{Pupil gender} 
\item{texp}{Teacher experience (years)} 
\item{const}{Constant intercept term}
\item{teachpop}{Teacher popularity} }
}
\source{
Hox, J. J. (2002) \emph{Multilevel analysis. Techniques and
applications.} Mahwah, NJ: Lawrence Erlbaum.
}
\description{
Hox pupil popularity data with some missing popularity scores
}
\details{
The original, complete dataset was generated by Joop Hox as an example of
well-behaved multilevel data set. The distributed data contains missing data
in pupil popularity.
}
\examples{

popmis[1:3,]

}
\keyword{datasets}
