% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.mads}
\alias{print.mads}
\title{Print a \code{mads} object}
\usage{
\method{print}{mads}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mads}}

\item{...}{Other parameters passed down to \code{print.default()}}
}
\value{
\code{NULL}
}
\description{
Print a \code{mads} object
}
\seealso{
\code{\link[=mads-class]{mads}}
}
