% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mira}
\alias{print.mira}
\alias{print.mice.anova}
\alias{print.mice.anova.summary}
\title{Print a \code{mira} object}
\usage{
\method{print}{mira}(x, ...)

\method{print}{mice.anova}(x, ...)

\method{print}{mice.anova.summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mice.anova}}

\item{\dots}{Other arguments}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Print a \code{mira} object

Print a \code{mice.anova} object

Print a \code{summary.mice.anova} object
}
\seealso{
\code{\link{mira}}

\code{\link{mipo}}

\code{\link{mipo}}
}
