% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.mids}
\alias{print.mids}
\alias{print.mira}
\alias{print.mipo}
\title{Print a \code{mids} object}
\usage{
\method{print}{mids}(x, ...)

\method{print}{mira}(x, ...)

\method{print}{mipo}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mids}, \code{mira} or \code{mipo}}

\item{...}{Other parameters passed down to \code{print.default()}}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Print a \code{mids} object

Print a \code{mira} object

Print a \code{mipo} object
}
\seealso{
\code{\link[=mids-class]{mids}}

\code{\link[=mira-class]{mira}}

\code{\link[=mipo-class]{mipo}}
}
