% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-main.R
\name{runMICE}
\alias{runMICE}
\title{Run a simulation of the MICE model}
\usage{
runMICE(groups, fleets, Mstarv = 0.3, Ystar = 0.4, delta = 0.9, T = 100,
  ndtPerYear = 4, verbose = TRUE)
}
\arguments{
\item{Mstarv}{Maximum starvation mortality}

\item{Ystar}{Optimal annual food ration per gram of biomass}

\item{delta}{Fraction of prey population available to predators}

\item{T}{Time horizon for the simulation (years).}

\item{LHT}{List with the life history parameters}

\item{B0}{Initial biomass for all species}

\item{predRange}{Minimum and maximum predator-prey ratios}

\item{dt}{Time step for simulation, in fractions of a year}
}
\value{
A list with the abundance (N), length (L) and biomass (B) of
all the species modeled.
}
\description{
Run a simulation of the MICE model
}
