% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{runMICE}
\alias{runMICE}
\title{Run a simulation of the MICE model}
\usage{
runMICE(LHT, B0, F = 0, predRange = c(0.1, 0.25), Mstarv = 0.3,
  Ystar = 0.4, delta = 0.9, dt = 1/4, T = 100/dt)
}
\arguments{
\item{LHT}{List with the life history parameters}

\item{B0}{Initial biomass for all species}

\item{predRange}{Minimum and maximum predator-prey ratios}

\item{Mstarv}{Maximum starvation mortality}

\item{Ystar}{Optimal annual food ration per gram of biomass}

\item{delta}{Fraction of prey population available to predators}

\item{dt}{Time step for simulation, in fractions of a year}

\item{T}{Time horizon for the simulation (years).}
}
\value{
A list with the abundance (N), length (L) and biomass (B) of
all the species modeled.
}
\description{
Run a simulation of the MICE model
}
