% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-main.R
\name{runMICE}
\alias{runMICE}
\title{Run a simulation of the MICE model}
\usage{
runMICE(groups, fleets, T, ndtPerYear = 4, Mstarv = 0.3, Ystar = 3.5,
  delta = 0.9, par = NULL, Fmult = 1, niter = 7, verbose = TRUE)
}
\arguments{
\item{groups}{A list containing the information to create the functional groups. See details.}

\item{fleets}{A list containing the information to create the fleets. See details.}

\item{ndtPerYear}{Number of time steps per year.}

\item{Mstarv}{Maximum starvation mortality.}

\item{Ystar}{Optimal annual food ration per gram of biomass.}

\item{delta}{Fraction of prey population available to predators.}

\item{par}{A list with parameters values. These parameters take precedence over 'groups' and 'fleets'.}

\item{Fmult}{Fishing multiplier. All the fishing mortalities are multiplied by this value.}

\item{niter}{Number of iterations for the calculation of predation mortality.}

\item{verbose}{Logical, should running messages be produced?}
}
\value{
A list with the abundance (N), length (L) and biomass (B) of
all the species modeled.
}
\description{
Run a simulation of the MICE model
}
