% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-main.R
\name{runMICE}
\alias{runMICE}
\title{Run a simulation of the MICE model}
\usage{
runMICE(groups, fleets, T, ndtPerYear = 4, Mstarv = 0.3, Ystar = 3.5,
  delta = 0.9, Fmult = 1, niter = 7, verbose = TRUE)
}
\arguments{
\item{T}{Time horizon for the simulation (years).}

\item{Mstarv}{Maximum starvation mortality}

\item{Ystar}{Optimal annual food ration per gram of biomass}

\item{delta}{Fraction of prey population available to predators}

\item{LHT}{List with the life history parameters}

\item{B0}{Initial biomass for all species}

\item{predRange}{Minimum and maximum predator-prey ratios}

\item{dt}{Time step for simulation, in fractions of a year}
}
\value{
A list with the abundance (N), length (L) and biomass (B) of
all the species modeled.
}
\description{
Run a simulation of the MICE model
}
