% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipo.r, R/summary.r
\name{summary.mipo}
\alias{summary.mipo}
\alias{summary.mira}
\alias{summary.mids}
\alias{summary.mads}
\alias{summary.mice.anova}
\title{Summary of a \code{mipo} object}
\usage{
\method{summary}{mipo}(object, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

\method{summary}{mira}(object, type = c("tidy", "glance", "summary"), ...)

\method{summary}{mids}(object, ...)

\method{summary}{mads}(object, ...)

\method{summary}{mice.anova}(object, ...)
}
\arguments{
\item{object}{A \code{mira} object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for logistic regression)}

\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}

\item{type}{A length-1 character vector indicating the 
type of summary. There are three choices: \code{type = "tidy"}
return the parameters estimates of each analyses as a data frame.
\code{type = "glance"} return the fit statistics of each analysis
as a data frame. \code{type = "summary"} returns a list of
length \code{m} with the analysis results. The default is 
\code{"tidy"}.}
}
\value{
A table containing summary statistis of the pooled analysis

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Summary of a \code{mipo} object

Summary of a \code{mira} object

Summary of a \code{mids} object

Summary of a \code{mads} object

Print a \code{mice.anova} object
}
\seealso{
\code{\link{mipo}}

\code{\link[=mira-class]{mira}}

\code{\link[=mids-class]{mids}}

\code{\link[=mads-class]{mads}}

\code{\link{mipo}}
}
