% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mira}
\alias{summary.mira}
\alias{summary.mids}
\alias{summary.mads}
\alias{summary.mice.anova}
\title{Summary of a `mira` object}
\usage{
\method{summary}{mira}(object, type = c("tidy", "glance", "summary"), ...)

\method{summary}{mids}(object, ...)

\method{summary}{mads}(object, ...)

\method{summary}{mice.anova}(object, ...)
}
\arguments{
\item{object}{A `mira` object}

\item{type}{A length-1 character vector indicating the
type of summary. There are three choices: `type = "tidy"`
return the parameters estimates of each analyses as a data frame.
`type = "glance"` return the fit statistics of each analysis
as a data frame. `type = "summary"` returns a list of
length `m` with the analysis results. The default is
`"tidy"`.}

\item{...}{Other parameters passed down to `print()` and `summary()`}
}
\value{
`NULL`

`NULL`

`NULL`

`NULL`
}
\description{
Summary of a `mira` object

Summary of a `mids` object

Summary of a `mads` object

Print a `mice.anova` object
}
\seealso{
[`mira()`][mira-class]

[`mids()`][mids-class]

[`mads()`][mads-class]

[mipo()]
}
