% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mira}
\alias{summary.mira}
\alias{summary.mice.anova}
\title{Summary of a \code{mira} object}
\usage{
\method{summary}{mira}(object, type = c("tidy", "glance", "summary"), dfcom = NULL, ...)

\method{summary}{mice.anova}(object, ...)
}
\arguments{
\item{object}{A \code{mira} object}

\item{type}{A length-1 character vector indicating the
type of summary. There are three choices: \code{type = "tidy"}
return the parameters estimates of each analyses as a data frame.
\code{type = "glance"} return the fit statistics of each analysis
as a data frame. \code{type = "summary"} returns a list of
length \code{m} with the analysis results. The default is
\code{"tidy"}.}

\item{dfcom}{Manually supplied degrees of freedom. For internal use by
\code{pool()}.}

\item{...}{Other parameters passed down to \code{print()} and \code{summary()}}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Summary of a \code{mira} object

Print a \code{mice.anova} object
}
\seealso{
\code{\link{mira}}

\code{\link{mipo}}
}
