\name{summary.mira}
\alias{summary.mids}
\alias{summary.mipo}
\alias{summary.mira}
\title{Summary of a \code{mira} object}
\usage{
\method{summary}{mira}(object, ...)

\method{summary}{mipo}(object, ...)

\method{summary}{mids}(object, ...)
}
\arguments{
  \item{object}{A \code{mira} object}

  \item{...}{Other parameters passed down to \code{print()}
  and \code{summary()}}
}
\value{
\code{NULL}

A table containing summary statistis of the pooled analysis

\code{NULL}
}
\description{
Summary of a \code{mira} object

Summary of a \code{mipo} object

Summary of a \code{mids} object
}
\seealso{
\code{\link[=mira-class]{mira}}

\code{\link[=mipo-class]{mipo}}

\code{\link[=mids-class]{mids}}
}

