% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.predictors.R
\name{trim.predictors}
\alias{trim.predictors}
\title{Trim predictors of the imputation model}
\usage{
trim.predictors(
  x,
  y,
  ry,
  trimmer = c("lars.filter", "remove.lindep", "none"),
  ...
)
}
\arguments{
\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{trimmer}{A character vector of length 1 specifying the name of the
trimming function. The default is "lars.filter". The other options are
"remove.lindep" and "none". The user can also specify a custom trimming
function.}

\item{...}{further arguments passed to the trimming function.}
}
\value{
A logical vector of length `ncol(x)` indicating which predictors
to keep.
}
\description{
The function `trim.predictors()` filters out predictors before imputation
of univariate incomplete variables. The function is called by
`mice:::sampler.univ()`. The function is a wrapper for the trimming
functions `lars.filter()`, `remove.lindep()`, and a user-specified
trimming function. The default method is `"lars.filter"`.
}
