% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\name{version}
\alias{version}
\title{Echoes the package version number}
\usage{
version(pkg = "mice")
}
\arguments{
\item{pkg}{A character vector with the package name.}
}
\value{
A character vector containing the package name, version number and
installed directory.
}
\description{
Echoes the package version number
}
\examples{
version()
version("base")
}
\author{
Stef van Buuren, Oct 2010
}
\keyword{misc}

